remove opengl_mobject.py and rename OpenGLMobject -> Mobject
