+++ manim/mobject/mobject.py
@@ -1,68 +1,44 @@
-"""Base classes for objects that can be displayed."""
-
-
-__all__ = ["Mobject", "Group", "override_animate"]
-
-
 import copy
 import itertools as it
-import operator as op
 import random
 import sys
-import types
-import warnings
-from functools import partialmethod, reduce
+from functools import partialmethod, wraps
 from math import ceil
-from pathlib import Path
-from typing import (
-    TYPE_CHECKING,
-    Callable,
-    Dict,
-    Iterable,
-    List,
-    Optional,
-    Sequence,
-    Tuple,
-    Type,
-    TypeVar,
-    Union,
-)
+from typing import Iterable, Optional, Sequence, Tuple, Union
 
+import moderngl
 import numpy as np
 from colour import Color
 
 from .. import config
 from ..constants import *
-from ..utils.color import (
-    BLACK,
-    WHITE,
-    YELLOW_C,
-    Colors,
-    color_gradient,
-    interpolate_color,
+from ..utils.bezier import integer_interpolate, interpolate
+from ..utils.color import *
+from ..utils.color import Colors
+from ..utils.config_ops import _Data, _Uniforms
+
+# from ..utils.iterables import batch_by_property
+from ..utils.iterables import (
+    batch_by_property,
+    list_update,
+    listify,
+    make_even,
+    resize_array,
+    resize_preserving_order,
+    resize_with_interpolation,
+    uniq_chain,
 )
-from ..utils.exceptions import MultiAnimationOverrideException
-from ..utils.iterables import list_update, remove_list_redundancies
 from ..utils.paths import straight_path
 from ..utils.simple_functions import get_parameters
-from ..utils.space_ops import angle_between_vectors, normalize, rotation_matrix
-from .opengl_compatibility import ConvertToOpenGL
-
-# TODO: Explain array_attrs
-
-Updater = Union[Callable[["Mobject"], None], Callable[["Mobject", float], None]]
-T = TypeVar("T", bound="Mobject")
-
-if TYPE_CHECKING:
-    from ..animation.animation import Animation
+from ..utils.space_ops import (
+    angle_between_vectors,
+    normalize,
+    rotation_matrix_transpose,
+)
 
 
 class Mobject:
     """Mathematical Object: base class for objects that can be displayed on screen.
-
-    There is a compatibility layer that allows for
-    getting and setting generic attributes with ``get_*``
-    and ``set_*`` methods. See :meth:`set` for more details.
 
     Attributes
     ----------
@@ -73,107 +49,115 @@ class Mobject:
 
         .. seealso::
 
-            :class:`~.VMobject`
+            :class:`~.OpenGLVMobject`
 
     """
 
-    animation_overrides = {}
+    shader_dtype = [
+        ("point", np.float32, (3,)),
+    ]
+    shader_folder = ""
+
+    # _Data and _Uniforms are set as class variables to tell manim how to handle setting/getting these attributes later.
+    points = _Data()
+    bounding_box = _Data()
+    rgbas = _Data()
+
+    is_fixed_in_frame = _Uniforms()
+    gloss = _Uniforms()
+    shadow = _Uniforms()
+
+    def __init__(
+        self,
+        color=WHITE,
+        opacity=1,
+        dim=3,  # TODO, get rid of this
+        # Lighting parameters
+        # Positive gloss up to 1 makes it reflect the light.
+        gloss=0.0,
+        # Positive shadow up to 1 makes a side opposite the light darker
+        shadow=0.0,
+        # For shaders
+        render_primitive=moderngl.TRIANGLES,
+        texture_paths=None,
+        depth_test=False,
+        # If true, the mobject will not get rotated according to camera position
+        is_fixed_in_frame=False,
+        # Must match in attributes of vert shader
+        # Event listener
+        listen_to_events=False,
+        model_matrix=None,
+        should_render=True,
+        **kwargs,
+    ):
+        # getattr in case data/uniforms are already defined in parent classes.
+        self.data = getattr(self, "data", {})
+        self.uniforms = getattr(self, "uniforms", {})
+
+        self.color = Color(color) if color else None
+        self.opacity = opacity
+        self.dim = dim  # TODO, get rid of this
+        # Lighting parameters
+        # Positive gloss up to 1 makes it reflect the light.
+        self.gloss = gloss
+        # Positive shadow up to 1 makes a side opposite the light darker
+        self.shadow = shadow
+        # For shaders
+        self.render_primitive = render_primitive
+        self.texture_paths = texture_paths
+        self.depth_test = depth_test
+        # If true, the mobject will not get rotated according to camera position
+        self.is_fixed_in_frame = float(is_fixed_in_frame)
+        # Must match in attributes of vert shader
+        # Event listener
+        self.listen_to_events = listen_to_events
+
+        self._submobjects = []
+        self.parents = []
+        self.parent = None
+        self.family = [self]
+        self.locked_data_keys = set()
+        self.needs_new_bounding_box = True
+        if model_matrix is None:
+            self.model_matrix = np.eye(4)
+        else:
+            self.model_matrix = model_matrix
+
+        self.init_data()
+        self.init_updaters()
+        # self.init_event_listners()
+        self.init_points()
+        self.init_colors()
+
+        self.shader_indices = None
+
+        if self.depth_test:
+            self.apply_depth_test()
+
+        self.should_render = should_render
 
     @classmethod
     def __init_subclass__(cls, **kwargs):
         super().__init_subclass__(**kwargs)
-
-        cls.animation_overrides: Dict[
-            Type["Animation"],
-            Callable[["Mobject"], "Animation"],
-        ] = {}
-        cls._add_intrinsic_animation_overrides()
         cls._original__init__ = cls.__init__
 
-    def __init__(self, color=WHITE, name=None, dim=3, target=None, z_index=0):
-        self.color = Color(color) if color else None
-        self.name = self.__class__.__name__ if name is None else name
-        self.dim = dim
-        self.target = target
-        self.z_index = z_index
-        self.point_hash = None
-        self.submobjects = []
-        self.updaters = []
-        self.updating_suspended = False
-        self.reset_points()
-        self.generate_points()
-        self.init_colors()
+    def __str__(self):
+        return self.__class__.__name__
 
-    @classmethod
-    def animation_override_for(
-        cls,
-        animation_class: Type["Animation"],
-    ) -> "Optional[Callable[[Mobject, ...], Animation]]":
-        """Returns the function defining a specific animation override for this class.
+    def __repr__(self):
+        return self.__class__.__name__
 
-        Parameters
-        ----------
-        animation_class
-            The animation class for which the override function should be returned.
+    def __sub__(self, other):
+        raise NotImplementedError
 
-        Returns
-        -------
-        Optional[Callable[[Mobject, ...], Animation]]
-            The function returning the override animation or ``None`` if no such animation
-            override is defined.
-        """
-        if animation_class in cls.animation_overrides:
-            return cls.animation_overrides[animation_class]
+    def __isub__(self, other):
+        raise NotImplementedError
 
-        return None
+    def __add__(self, mobject):
+        raise NotImplementedError
 
-    @classmethod
-    def _add_intrinsic_animation_overrides(cls):
-        """Initializes animation overrides marked with the :func:`~.override_animation`
-        decorator.
-        """
-        for method_name in dir(cls):
-            # Ignore dunder methods
-            if method_name.startswith("__"):
-                continue
-
-            method = getattr(cls, method_name)
-            if hasattr(method, "_override_animation"):
-                animation_class = method._override_animation
-                cls.add_animation_override(animation_class, method)
-
-    @classmethod
-    def add_animation_override(
-        cls,
-        animation_class: Type["Animation"],
-        override_func: "Callable[[Mobject, ...], Animation]",
-    ):
-        """Add an animation override.
-
-        This does not apply to subclasses.
-
-        Parameters
-        ----------
-        animation_class
-            The animation type to be overridden
-        override_func
-            The function returning an animation replacing the default animation. It gets
-            passed the parameters given to the animnation constructor.
-
-        Raises
-        ------
-        MultiAnimationOverrideException
-            If the overridden animation was already overridden.
-        """
-        if animation_class not in cls.animation_overrides:
-            cls.animation_overrides[animation_class] = override_func
-        else:
-            raise MultiAnimationOverrideException(
-                f"The animation {animation_class.__name__} for "
-                f"{cls.__name__} is overridden by more than one method: "
-                f"{cls.animation_overrides[animation_class].__qualname__} and "
-                f"{override_func.__qualname__}.",
-            )
+    def __iadd__(self, mobject):
+        raise NotImplementedError
 
     @classmethod
     def set_default(cls, **kwargs):
@@ -223,21 +207,74 @@ class Mobject:
         else:
             cls.__init__ = cls._original__init__
 
-    @property
-    def animate(self):
-        """Used to animate the application of any method of :code:`self`.
+    def init_data(self):
+        """Initializes the ``points``, ``bounding_box`` and ``rgbas`` attributes and groups them into self.data.
+        Subclasses can inherit and overwrite this method to extend `self.data`."""
+        self.points = np.zeros((0, 3))
+        self.bounding_box = np.zeros((3, 3))
+        self.rgbas = np.zeros((1, 4))
 
-        Any method called on :code:`animate` is converted to an animation of applying
-        that method on the mobject itself.
+    def init_colors(self):
+        """Initializes the colors.
 
-        For example, :code:`square.set_fill(WHITE)` sets the fill color of a square,
-        while :code:`sqaure.animate.set_fill(WHITE)` animates this action.
+        Gets called upon creation"""
+        self.set_color(self.color, self.opacity)
 
-        Multiple methods can be put in a single animation once via chaining:
+    def init_points(self):
+        """Initializes :attr:`points` and therefore the shape.
 
+        Gets called upon creation. This is an empty method that can be implemented by
+        subclasses."""
+        # Typically implemented in subclass, unless purposefully left blank
+        pass
+
+    def set(self, **kwargs) -> "Mobject":
+        """Sets attributes.
+
+        Mainly to be used along with :attr:`animate` to
+        animate setting attributes.
+
+        Examples
+        --------
         ::
 
-            self.play(my_mobject.animate.shift(RIGHT).rotate(PI))
+            >>> mob = Mobject()
+            >>> mob.set(foo=0)
+            Mobject
+            >>> mob.foo
+            0
+
+        Parameters
+        ----------
+        **kwargs
+            The attributes and corresponding values to set.
+
+        Returns
+        -------
+        :class:`Mobject`
+            ``self``
+
+
+        """
+
+        for attr, value in kwargs.items():
+            setattr(self, attr, value)
+
+        return self
+
+    def set_data(self, data):
+        for key in data:
+            self.data[key] = data[key].copy()
+        return self
+
+    def set_uniforms(self, uniforms):
+        for key in uniforms:
+            self.uniforms[key] = uniforms[key]  # Copy?
+        return self
+
+    @property
+    def animate(self):
+        """Used to animate the application of a method.
 
         .. warning::
 
@@ -249,7 +286,9 @@ class Mobject:
 
                 self.play(my_mobject.animate.shift(RIGHT), my_mobject.animate.rotate(PI))
 
-            make use of method chaining.
+            make use of method chaining for ``animate``, meaning::
+
+                self.play(my_mobject.animate.shift(RIGHT).rotate(PI))
 
         Keyword arguments that can be passed to :meth:`.Scene.play` can be passed
         directly after accessing ``.animate``, like so::
@@ -321,42 +360,301 @@ class Mobject:
         """
         return _AnimationBuilder(self)
 
-    def __deepcopy__(self, clone_from_id):
-        cls = self.__class__
-        result = cls.__new__(cls)
-        clone_from_id[id(self)] = result
-        for k, v in self.__dict__.items():
-            setattr(result, k, copy.deepcopy(v, clone_from_id))
-        result.original_id = str(id(self))
-        return result
+    @property
+    def width(self):
+        """The width of the mobject.
 
-    def __repr__(self):
-        if config["renderer"] == "opengl":
-            return super().__repr__()
+        Returns
+        -------
+        :class:`float`
+
+        Examples
+        --------
+        .. manim:: WidthExample
+
+            class WidthExample(Scene):
+                def construct(self):
+                    decimal = DecimalNumber().to_edge(UP)
+                    rect = Rectangle(color=BLUE)
+                    rect_copy = rect.copy().set_stroke(GRAY, opacity=0.5)
+
+                    decimal.add_updater(lambda d: d.set_value(rect.width))
+
+                    self.add(rect_copy, rect, decimal)
+                    self.play(rect.animate.set(width=7))
+                    self.wait()
+
+        See also
+        --------
+        :meth:`length_over_dim`
+
+        """
+
+        # Get the length across the X dimension
+        return self.length_over_dim(0)
+
+    # Only these methods should directly affect points
+    @width.setter
+    def width(self, value):
+        self.rescale_to_fit(value, 0, stretch=False)
+
+    @property
+    def height(self):
+        """The height of the mobject.
+
+        Returns
+        -------
+        :class:`float`
+
+        Examples
+        --------
+        .. manim:: HeightExample
+
+            class HeightExample(Scene):
+                def construct(self):
+                    decimal = DecimalNumber().to_edge(UP)
+                    rect = Rectangle(color=BLUE)
+                    rect_copy = rect.copy().set_stroke(GRAY, opacity=0.5)
+
+                    decimal.add_updater(lambda d: d.set_value(rect.height))
+
+                    self.add(rect_copy, rect, decimal)
+                    self.play(rect.animate.set(height=5))
+                    self.wait()
+
+        See also
+        --------
+        :meth:`length_over_dim`
+
+        """
+
+        # Get the length across the Y dimension
+        return self.length_over_dim(1)
+
+    @height.setter
+    def height(self, value):
+        self.rescale_to_fit(value, 1, stretch=False)
+
+    @property
+    def depth(self):
+        """The depth of the mobject.
+
+        Returns
+        -------
+        :class:`float`
+
+        See also
+        --------
+        :meth:`length_over_dim`
+
+        """
+
+        # Get the length across the Z dimension
+        return self.length_over_dim(2)
+
+    @depth.setter
+    def depth(self, value):
+        self.rescale_to_fit(value, 2, stretch=False)
+
+    def resize_points(self, new_length, resize_func=resize_array):
+        if new_length != len(self.points):
+            self.points = resize_func(self.points, new_length)
+        self.refresh_bounding_box()
+        return self
+
+    def set_points(self, points):
+        if len(points) == len(self.points):
+            self.points[:] = points
+        elif isinstance(points, np.ndarray):
+            self.points = points.copy()
         else:
-            return str(self.name)
+            self.points = np.array(points)
+        self.refresh_bounding_box()
+        return self
 
-    def reset_points(self):
-        """Sets :attr:`points` to be an empty array."""
-        self.points = np.zeros((0, self.dim))
+    def apply_over_attr_arrays(self, func):
+        for attr in self.get_array_attrs():
+            setattr(self, attr, func(getattr(self, attr)))
+        return self
 
-    def init_colors(self):
-        """Initializes the colors.
+    def append_points(self, new_points):
+        self.points = np.vstack([self.points, new_points])
+        self.refresh_bounding_box()
+        return self
 
-        Gets called upon creation. This is an empty method that can be implemented by
-        subclasses.
+    def reverse_points(self):
+        for mob in self.get_family():
+            for key in mob.data:
+                mob.data[key] = mob.data[key][::-1]
+        return self
+
+    def get_midpoint(self) -> np.ndarray:
+        """Get coordinates of the middle of the path that forms the  :class:`~.Mobject`.
+
+        Examples
+        --------
+
+        .. manim:: AngleMidPoint
+            :save_last_frame:
+
+            class AngleMidPoint(Scene):
+                def construct(self):
+                    line1 = Line(ORIGIN, 2*RIGHT)
+                    line2 = Line(ORIGIN, 2*RIGHT).rotate_about_origin(80*DEGREES)
+
+                    a = Angle(line1, line2, radius=1.5, other_angle=False)
+                    d = Dot(a.get_midpoint()).set_color(RED)
+
+                    self.add(line1, line2, a, d)
+                    self.wait()
+
         """
-        pass
+        return self.point_from_proportion(0.5)
 
-    def generate_points(self):
-        """Initializes :attr:`points` and therefore the shape.
+    def apply_points_function(
+        self,
+        func,
+        about_point=None,
+        about_edge=ORIGIN,
+        works_on_bounding_box=False,
+    ):
+        if about_point is None and about_edge is not None:
+            about_point = self.get_bounding_box_point(about_edge)
 
-        Gets called upon creation. This is an empty method that can be implemented by
-        subclasses.
+        for mob in self.get_family():
+            arrs = []
+            if mob.has_points():
+                arrs.append(mob.points)
+            if works_on_bounding_box:
+                arrs.append(mob.get_bounding_box())
+
+            for arr in arrs:
+                if about_point is None:
+                    arr[:] = func(arr)
+                else:
+                    arr[:] = func(arr - about_point) + about_point
+
+        if not works_on_bounding_box:
+            self.refresh_bounding_box(recurse_down=True)
+        else:
+            for parent in self.parents:
+                parent.refresh_bounding_box()
+        return self
+
+    # Others related to points
+
+    def match_points(self, mobject):
+        """Edit points, positions, and submobjects to be identical
+        to another :class:`~.Mobject`, while keeping the style unchanged.
+
+        Examples
+        --------
+        .. manim:: MatchPointsScene
+
+            class MatchPointsScene(Scene):
+                def construct(self):
+                    circ = Circle(fill_color=RED, fill_opacity=0.8)
+                    square = Square(fill_color=BLUE, fill_opacity=0.2)
+                    self.add(circ)
+                    self.wait(0.5)
+                    self.play(circ.animate.match_points(square))
+                    self.wait(0.5)
         """
-        pass
+        self.set_points(mobject.points)
 
-    def add(self, *mobjects: "Mobject"):
+    def clear_points(self):
+        self.resize_points(0)
+
+    def get_num_points(self):
+        return len(self.points)
+
+    def get_all_points(self):
+        if self.submobjects:
+            return np.vstack([sm.points for sm in self.get_family()])
+        else:
+            return self.points
+
+    def has_points(self):
+        return self.get_num_points() > 0
+
+    def get_bounding_box(self):
+        if self.needs_new_bounding_box:
+            self.bounding_box = self.compute_bounding_box()
+            self.needs_new_bounding_box = False
+        return self.bounding_box
+
+    def compute_bounding_box(self):
+        all_points = np.vstack(
+            [
+                self.points,
+                *(
+                    mob.get_bounding_box()
+                    for mob in self.get_family()[1:]
+                    if mob.has_points()
+                ),
+            ],
+        )
+        if len(all_points) == 0:
+            return np.zeros((3, self.dim))
+        else:
+            # Lower left and upper right corners
+            mins = all_points.min(0)
+            maxs = all_points.max(0)
+            mids = (mins + maxs) / 2
+            return np.array([mins, mids, maxs])
+
+    def refresh_bounding_box(self, recurse_down=False, recurse_up=True):
+        for mob in self.get_family(recurse_down):
+            mob.needs_new_bounding_box = True
+        if recurse_up:
+            for parent in self.parents:
+                parent.refresh_bounding_box()
+        return self
+
+    def is_point_touching(self, point, buff=MED_SMALL_BUFF):
+        bb = self.get_bounding_box()
+        mins = bb[0] - buff
+        maxs = bb[2] + buff
+        return (point >= mins).all() and (point <= maxs).all()
+
+    # Family matters
+
+    def __getitem__(self, value):
+        if isinstance(value, slice):
+            GroupClass = self.get_group_class()
+            return GroupClass(*self.split().__getitem__(value))
+        return self.split().__getitem__(value)
+
+    def __iter__(self):
+        return iter(self.split())
+
+    def __len__(self):
+        return len(self.split())
+
+    def split(self):
+        return self.submobjects
+
+    def assemble_family(self):
+        sub_families = (sm.get_family() for sm in self.submobjects)
+        self.family = [self, *uniq_chain(*sub_families)]
+        self.refresh_has_updater_status()
+        self.refresh_bounding_box()
+        for parent in self.parents:
+            parent.assemble_family()
+        return self
+
+    def get_family(self, recurse=True):
+        if recurse:
+            return self.family
+        else:
+            return [self]
+
+    def family_members_with_points(self):
+        return [m for m in self.get_family() if m.has_points()]
+
+    def add(
+        self, *mobjects: "Mobject", update_parent: bool = False
+    ) -> "Mobject":
         """Add mobjects as submobjects.
 
         The mobjects are added to :attr:`submobjects`.
@@ -415,21 +713,60 @@ class Mobject:
             ValueError: Mobject cannot contain self
 
         """
-        for m in mobjects:
-            if not isinstance(m, Mobject):
+        if update_parent:
+            assert len(mobjects) == 1, "Can't set multiple parents."
+            mobjects[0].parent = self
+
+        if self in mobjects:
+            raise ValueError("Mobject cannot contain self")
+        for mobject in mobjects:
+            if not isinstance(mobject, Mobject):
                 raise TypeError("All submobjects must be of type Mobject")
-            if m is self:
-                raise ValueError("Mobject cannot contain self")
-        self.submobjects = list_update(self.submobjects, mobjects)
+            if mobject not in self.submobjects:
+                self.submobjects.append(mobject)
+            if self not in mobject.parents:
+                mobject.parents.append(self)
+        self.assemble_family()
         return self
 
-    def __add__(self, mobject):
-        raise NotImplementedError
+    def remove(
+        self, *mobjects: "Mobject", update_parent: bool = False
+    ) -> "Mobject":
+        """Remove :attr:`submobjects`.
 
-    def __iadd__(self, mobject):
-        raise NotImplementedError
+        The mobjects are removed from :attr:`submobjects`, if they exist.
 
-    def add_to_back(self, *mobjects: "Mobject"):
+        Subclasses of mobject may implement ``-`` and ``-=`` dunder methods.
+
+        Parameters
+        ----------
+        mobjects
+            The mobjects to remove.
+
+        Returns
+        -------
+        :class:`Mobject`
+            ``self``
+
+        See Also
+        --------
+        :meth:`add`
+
+        """
+        if update_parent:
+            assert len(mobjects) == 1, "Can't remove multiple parents."
+            mobjects[0].parent = None
+
+        for mobject in mobjects:
+            if mobject in self.submobjects:
+                self.submobjects.remove(mobject)
+            if self in mobject.parents:
+                mobject.parents.remove(self)
+        self.assemble_family()
+        return self
+
+    def add_to_back(self, *mobjects: "Mobject") -> "Mobject":
+        # NOTE: is the note true Mobjects?
         """Add all passed mobjects to the back of the submobjects.
 
         If :attr:`submobjects` already contains the given mobjects, they just get moved
@@ -473,1650 +810,47 @@ class Mobject:
         :meth:`add`
 
         """
-        if self in mobjects:
-            raise ValueError("A mobject shouldn't contain itself")
-
-        for mobject in mobjects:
-            if not isinstance(mobject, Mobject):
-                raise TypeError("All submobjects must be of type Mobject")
-
-        self.remove(*mobjects)
-        # dict.fromkeys() removes duplicates while maintaining order
-        self.submobjects = list(dict.fromkeys(mobjects)) + self.submobjects
+        self.submobjects = list_update(mobjects, self.submobjects)
         return self
 
-    def remove(self, *mobjects: "Mobject"):
-        """Remove :attr:`submobjects`.
-
-        The mobjects are removed from :attr:`submobjects`, if they exist.
-
-        Subclasses of mobject may implement ``-`` and ``-=`` dunder methods.
-
-        Parameters
-        ----------
-        mobjects
-            The mobjects to remove.
-
-        Returns
-        -------
-        :class:`Mobject`
-            ``self``
-
-        See Also
-        --------
-        :meth:`add`
-
-        """
-        for mobject in mobjects:
-            if mobject in self.submobjects:
-                self.submobjects.remove(mobject)
+    def replace_submobject(self, index, new_submob):
+        old_submob = self.submobjects[index]
+        if self in old_submob.parents:
+            old_submob.parents.remove(self)
+        self.submobjects[index] = new_submob
+        self.assemble_family()
         return self
 
-    def __sub__(self, other):
-        raise NotImplementedError
-
-    def __isub__(self, other):
-        raise NotImplementedError
-
-    def set(self, **kwargs):
-        """Sets attributes.
-
-        I.e. ``my_mobject.set(foo=1)`` applies ``my_mobject.foo = 1``.
-
-        This is a convenience to be used along with :attr:`animate` to
-        animate setting attributes.
-
-        In addition to this method, there is a compatibility
-        layer that allows ``get_*`` and ``set_*`` methods to
-        get and set generic attributes. For instance::
-
-            >>> mob = Mobject()
-            >>> mob.set_foo(0)
-            Mobject
-            >>> mob.get_foo()
-            0
-            >>> mob.foo
-            0
-
-        This compatibility layer does not interfere with any
-        ``get_*`` or ``set_*`` methods that are explicitly
-        defined.
-
-        .. warning::
-
-            This compatibility layer is for backwards compatibility
-            and is not guaranteed to stay around. Where applicable,
-            please prefer getting/setting attributes normally or with
-            the :meth:`set` method.
-
-        Parameters
-        ----------
-        **kwargs
-            The attributes and corresponding values to set.
-
-        Returns
-        -------
-        :class:`Mobject`
-            ``self``
-
-        Examples
-        --------
-        ::
-
-            >>> mob = Mobject()
-            >>> mob.set(foo=0)
-            Mobject
-            >>> mob.foo
-            0
-        """
-
-        for attr, value in kwargs.items():
-            setattr(self, attr, value)
-
-        return self
-
-    def __getattr__(self, attr):
-        # Add automatic compatibility layer
-        # between properties and get_* and set_*
-        # methods.
-        #
-        # In python 3.9+ we could change this
-        # logic to use str.remove_prefix instead.
-
-        if attr.startswith("get_"):
-            # Remove the "get_" prefix
-            to_get = attr[4:]
-
-            def getter(self):
-                warnings.warn(
-                    "This method is not guaranteed to stay around. Please prefer "
-                    "getting the attribute normally.",
-                    DeprecationWarning,
-                    stacklevel=2,
-                )
-
-                return getattr(self, to_get)
-
-            # Return a bound method
-            return types.MethodType(getter, self)
-
-        if attr.startswith("set_"):
-            # Remove the "set_" prefix
-            to_set = attr[4:]
-
-            def setter(self, value):
-                warnings.warn(
-                    "This method is not guaranteed to stay around. Please prefer "
-                    "setting the attribute normally or with Mobject.set().",
-                    DeprecationWarning,
-                    stacklevel=2,
-                )
-
-                setattr(self, to_set, value)
-
-                return self
-
-            # Return a bound method
-            return types.MethodType(setter, self)
-
-        # Unhandled attribute, therefore error
-        raise AttributeError(f"{type(self).__name__} object has no attribute '{attr}'")
-
-    @property
-    def width(self):
-        """The width of the mobject.
-
-        Returns
-        -------
-        :class:`float`
-
-        Examples
-        --------
-        .. manim:: WidthExample
-
-            class WidthExample(Scene):
-                def construct(self):
-                    decimal = DecimalNumber().to_edge(UP)
-                    rect = Rectangle(color=BLUE)
-                    rect_copy = rect.copy().set_stroke(GRAY, opacity=0.5)
-
-                    decimal.add_updater(lambda d: d.set_value(rect.width))
-
-                    self.add(rect_copy, rect, decimal)
-                    self.play(rect.animate.set(width=7))
-                    self.wait()
-
-        See also
-        --------
-        :meth:`length_over_dim`
-
-        """
-
-        # Get the length across the X dimension
-        return self.length_over_dim(0)
-
-    @width.setter
-    def width(self, value):
-        self.scale_to_fit_width(value)
-
-    @property
-    def height(self):
-        """The height of the mobject.
-
-        Returns
-        -------
-        :class:`float`
-
-        Examples
-        --------
-        .. manim:: HeightExample
-
-            class HeightExample(Scene):
-                def construct(self):
-                    decimal = DecimalNumber().to_edge(UP)
-                    rect = Rectangle(color=BLUE)
-                    rect_copy = rect.copy().set_stroke(GRAY, opacity=0.5)
-
-                    decimal.add_updater(lambda d: d.set_value(rect.height))
-
-                    self.add(rect_copy, rect, decimal)
-                    self.play(rect.animate.set(height=5))
-                    self.wait()
-
-        See also
-        --------
-        :meth:`length_over_dim`
-
-        """
-
-        # Get the length across the Y dimension
-        return self.length_over_dim(1)
-
-    @height.setter
-    def height(self, value):
-        self.scale_to_fit_height(value)
-
-    @property
-    def depth(self):
-        """The depth of the mobject.
-
-        Returns
-        -------
-        :class:`float`
-
-        See also
-        --------
-        :meth:`length_over_dim`
-
-        """
-
-        # Get the length across the Z dimension
-        return self.length_over_dim(2)
-
-    @depth.setter
-    def depth(self, value):
-        self.scale_to_fit_depth(value)
-
-    def get_array_attrs(self):
-        return ["points"]
-
-    def apply_over_attr_arrays(self, func):
-        for attr in self.get_array_attrs():
-            setattr(self, attr, func(getattr(self, attr)))
-        return self
-
-    # Displaying
-
-    def get_image(self, camera=None):
-        if camera is None:
-            from ..camera.camera import Camera
-
-            camera = Camera()
-        camera.capture_mobject(self)
-        return camera.get_image()
-
-    def show(self, camera=None):
-        self.get_image(camera=camera).show()
-
-    def save_image(self, name=None):
-        """Saves an image of only this :class:`Mobject` at its position to a png
-        file."""
-        self.get_image().save(
-            Path(config.get_dir("video_dir")).joinpath((name or str(self)) + ".png"),
-        )
-
-    def copy(self: T) -> T:
-        """Create and return an identical copy of the :class:`Mobject` including all
-        :attr:`submobjects`.
-
-        Returns
-        -------
-        :class:`Mobject`
-            The copy.
-
-        Note
-        ----
-        The clone is initially not visible in the Scene, even if the original was.
-        """
-        return copy.deepcopy(self)
-
-    def generate_target(self, use_deepcopy=False):
-        self.target = None  # Prevent unbounded linear recursion
-        if use_deepcopy:
-            self.target = copy.deepcopy(self)
-        else:
-            self.target = self.copy()
-        return self.target
-
-    # Updating
-
-    def update(self, dt: float = 0, recursive: bool = True):
-        """Apply all updaters.
-
-        Does nothing if updating is suspended.
-
-        Parameters
-        ----------
-        dt
-            The parameter ``dt`` to pass to the update functions. Usually this is the
-            time in seconds since the last call of ``update``.
-        recursive
-            Whether to recursively update all submobjects.
-
-        Returns
-        -------
-        :class:`Mobject`
-            ``self``
-
-        See Also
-        --------
-        :meth:`add_updater`
-        :meth:`get_updaters`
-
-        """
-        if self.updating_suspended:
-            return self
-        for updater in self.updaters:
-            parameters = get_parameters(updater)
-            if "dt" in parameters:
-                updater(self, dt)
-            else:
-                updater(self)
-        if recursive:
-            for submob in self.submobjects:
-                submob.update(dt, recursive)
-        return self
-
-    def get_time_based_updaters(self) -> List[Updater]:
-        """Return all updaters using the ``dt`` parameter.
-
-        The updaters use this parameter as the input for difference in time.
-
-        Returns
-        -------
-        List[:class:`Callable`]
-            The list of time based updaters.
-
-        See Also
-        --------
-        :meth:`get_updaters`
-        :meth:`has_time_based_updater`
-
-        """
-        return [updater for updater in self.updaters if "dt" in get_parameters(updater)]
-
-    def has_time_based_updater(self) -> bool:
-        """Test if ``self`` has a time based updater.
-
-        Returns
-        -------
-        class:`bool`
-            ``True`` if at least one updater uses the ``dt`` parameter, ``False``
-            otherwise.
-
-        See Also
-        --------
-        :meth:`get_time_based_updaters`
-
-        """
-        return any("dt" in get_parameters(updater) for updater in self.updaters)
-
-    def get_updaters(self) -> List[Updater]:
-        """Return all updaters.
-
-        Returns
-        -------
-        List[:class:`Callable`]
-            The list of updaters.
-
-        See Also
-        --------
-        :meth:`add_updater`
-        :meth:`get_time_based_updaters`
-
-        """
-        return self.updaters
-
-    def get_family_updaters(self):
-        return list(it.chain(*(sm.get_updaters() for sm in self.get_family())))
-
-    def add_updater(
-        self,
-        update_function: Updater,
-        index: Optional[int] = None,
-        call_updater: bool = False,
-    ):
-        """Add an update function to this mobject.
-
-        Update functions, or updaters in short, are functions that are applied to the
-        Mobject in every frame.
-
-        Parameters
-        ----------
-        update_function
-            The update function to be added.
-            Whenever :meth:`update` is called, this update function gets called using
-            ``self`` as the first parameter.
-            The updater can have a second parameter ``dt``. If it uses this parameter,
-            it gets called using a second value ``dt``, usually representing the time
-            in seconds since the last call of :meth:`update`.
-        index
-            The index at which the new updater should be added in ``self.updaters``.
-            In case ``index`` is ``None`` the updater will be added at the end.
-        call_updater
-            Whether or not to call the updater initially. If ``True``, the updater will
-            be called using ``dt=0``.
-
-        Returns
-        -------
-        :class:`Mobject`
-            ``self``
-
-        Examples
-        --------
-        .. manim:: NextToUpdater
-
-            class NextToUpdater(Scene):
-                def construct(self):
-                    def dot_position(mobject):
-                        mobject.set_value(dot.get_center()[0])
-                        mobject.next_to(dot)
-
-                    dot = Dot(RIGHT*3)
-                    label = DecimalNumber()
-                    label.add_updater(dot_position)
-                    self.add(dot, label)
-
-                    self.play(Rotating(dot, about_point=ORIGIN, angle=TAU, run_time=TAU, rate_func=linear))
-
-        .. manim:: DtUpdater
-
-            class DtUpdater(Scene):
-                def construct(self):
-                    line = Square()
-
-                    #Let the line rotate 90° per second
-                    line.add_updater(lambda mobject, dt: mobject.rotate(dt*90*DEGREES))
-                    self.add(line)
-                    self.wait(2)
-
-        See also
-        --------
-        :meth:`get_updaters`
-        :meth:`remove_updater`
-        :class:`~.UpdateFromFunc`
-        """
-
-        if index is None:
-            self.updaters.append(update_function)
-        else:
-            self.updaters.insert(index, update_function)
-        if call_updater:
-            update_function(self, 0)
-        return self
-
-    def remove_updater(self, update_function: Updater):
-        """Remove an updater.
-
-        If the same updater is applied multiple times, every instance gets removed.
-
-        Parameters
-        ----------
-        update_function
-            The update function to be removed.
-
-
-        Returns
-        -------
-        :class:`Mobject`
-            ``self``
-
-        See also
-        --------
-        :meth:`clear_updaters`
-        :meth:`add_updater`
-        :meth:`get_updaters`
-
-        """
-        while update_function in self.updaters:
-            self.updaters.remove(update_function)
-        return self
-
-    def clear_updaters(self, recursive: bool = True):
-        """Remove every updater.
+    def invert(self, recursive=False):
+        """Inverts the list of :attr:`submobjects`.
 
         Parameters
         ----------
         recursive
-            Whether to recursively call ``clear_updaters`` on all submobjects.
+            If ``True``, all submobject lists of this mobject's family are inverted.
 
-        Returns
-        -------
-        :class:`Mobject`
-            ``self``
-
-        See also
+        Examples
         --------
-        :meth:`remove_updater`
-        :meth:`add_updater`
-        :meth:`get_updaters`
 
+        .. manim:: InvertSumobjectsExample
+
+            class InvertSumobjectsExample(Scene):
+                def construct(self):
+                    s = VGroup(*[Dot().shift(i*0.1*RIGHT) for i in range(-20,20)])
+                    s2 = s.copy()
+                    s2.invert()
+                    s2.shift(DOWN)
+                    self.play(Write(s), Write(s2))
         """
-        self.updaters = []
         if recursive:
             for submob in self.submobjects:
-                submob.clear_updaters()
-        return self
+                submob.invert(recursive=True)
+        list.reverse(self.submobjects)
+        self.assemble_family()
 
-    def match_updaters(self, mobject: "Mobject"):
-        """Match the updaters of the given mobject.
+    # Submobject organization
 
-        Parameters
-        ----------
-        mobject
-            The mobject whose updaters get matched.
-
-        Returns
-        -------
-        :class:`Mobject`
-            ``self``
-
-        Note
-        ----
-        All updaters from submobjects are removed, but only updaters of the given
-        mobject are matched, not those of it's submobjects.
-
-        See also
-        --------
-        :meth:`add_updater`
-        :meth:`clear_updaters`
-
-        """
-
-        self.clear_updaters()
-        for updater in mobject.get_updaters():
-            self.add_updater(updater)
-        return self
-
-    def suspend_updating(self, recursive: bool = True):
-        """Disable updating from updaters and animations.
-
-
-        Parameters
-        ----------
-        recursive
-            Whether to recursively suspend updating on all submobjects.
-
-        Returns
-        -------
-        :class:`Mobject`
-            ``self``
-
-        See also
-        --------
-        :meth:`resume_updating`
-        :meth:`add_updater`
-
-        """
-
-        self.updating_suspended = True
-        if recursive:
-            for submob in self.submobjects:
-                submob.suspend_updating(recursive)
-        return self
-
-    def resume_updating(self, recursive: bool = True):
-        """Enable updating from updaters and animations.
-
-        Parameters
-        ----------
-        recursive
-            Whether to recursively enable updating on all submobjects.
-
-        Returns
-        -------
-        :class:`Mobject`
-            ``self``
-
-        See also
-        --------
-        :meth:`suspend_updating`
-        :meth:`add_updater`
-
-        """
-        self.updating_suspended = False
-        if recursive:
-            for submob in self.submobjects:
-                submob.resume_updating(recursive)
-        self.update(dt=0, recursive=recursive)
-        return self
-
-    # Transforming operations
-
-    def apply_to_family(self, func: Callable[["Mobject"], None]):
-        """Apply a function to ``self`` and every submobject with points recursively.
-
-        Parameters
-        ----------
-        func
-            The function to apply to each mobject. ``func`` gets passed the respective
-            (sub)mobject as parameter.
-
-        Returns
-        -------
-        :class:`Mobject`
-            ``self``
-
-        See also
-        --------
-        :meth:`family_members_with_points`
-
-        """
-        for mob in self.family_members_with_points():
-            func(mob)
-
-    def shift(self, *vectors: np.ndarray):
-        """Shift by the given vectors.
-
-        Parameters
-        ----------
-        vectors
-            Vectors to shift by. If multiple vectors are given, they are added
-            together.
-
-        Returns
-        -------
-        :class:`Mobject`
-            ``self``
-
-        See also
-        --------
-        :meth:`move_to`
-        """
-
-        total_vector = reduce(op.add, vectors)
-        for mob in self.family_members_with_points():
-            mob.points = mob.points.astype("float")
-            mob.points += total_vector
-
-        return self
-
-    def scale(self, scale_factor: float, **kwargs):
-        r"""Scale the size by a factor.
-
-        Default behavior is to scale about the center of the mobject.
-
-        Parameters
-        ----------
-        scale_factor
-            The scaling factor :math:`\alpha`. If :math:`0 < |\alpha| < 1`, the mobject
-            will shrink, and for :math:`|\alpha| > 1` it will grow. Furthermore,
-            if :math:`\alpha < 0`, the mobject is also flipped.
-        kwargs
-            Additional keyword arguments passed to
-            :meth:`apply_points_function_about_point`.
-
-        Returns
-        -------
-        :class:`Mobject`
-            ``self``
-
-        Examples
-        --------
-
-        .. manim:: MobjectScaleExample
-            :save_last_frame:
-
-            class MobjectScaleExample(Scene):
-                def construct(self):
-                    f1 = Text("F")
-                    f2 = Text("F").scale(2)
-                    f3 = Text("F").scale(0.5)
-                    f4 = Text("F").scale(-1)
-
-                    vgroup = VGroup(f1, f2, f3, f4).arrange(6 * RIGHT)
-                    self.add(vgroup)
-
-        See also
-        --------
-        :meth:`move_to`
-
-        """
-        self.apply_points_function_about_point(
-            lambda points: scale_factor * points, **kwargs
-        )
-        return self
-
-    def rotate_about_origin(self, angle, axis=OUT, axes=[]):
-        """Rotates the :class:`~.Mobject` about the ORIGIN, which is at [0,0,0]."""
-        return self.rotate(angle, axis, about_point=ORIGIN)
-
-    def rotate(
-        self,
-        angle,
-        axis=OUT,
-        about_point: Optional[Sequence[float]] = None,
-        **kwargs,
-    ):
-        """Rotates the :class:`~.Mobject` about a certain point."""
-        rot_matrix = rotation_matrix(angle, axis)
-        self.apply_points_function_about_point(
-            lambda points: np.dot(points, rot_matrix.T), about_point, **kwargs
-        )
-        return self
-
-    def flip(self, axis=UP, **kwargs):
-        """Flips/Mirrors an mobject about its center.
-
-        Examples
-        --------
-
-        .. manim:: FlipExample
-            :save_last_frame:
-
-            class FlipExample(Scene):
-                def construct(self):
-                    s= Line(LEFT, RIGHT+UP).shift(4*LEFT)
-                    self.add(s)
-                    s2= s.copy().flip()
-                    self.add(s2)
-
-        """
-        return self.rotate(TAU / 2, axis, **kwargs)
-
-    def stretch(self, factor, dim, **kwargs):
-        def func(points):
-            points[:, dim] *= factor
-            return points
-
-        self.apply_points_function_about_point(func, **kwargs)
-        return self
-
-    def apply_function(self, function, **kwargs):
-        # Default to applying matrix about the origin, not mobjects center
-        if len(kwargs) == 0:
-            kwargs["about_point"] = ORIGIN
-        self.apply_points_function_about_point(
-            lambda points: np.apply_along_axis(function, 1, points), **kwargs
-        )
-        return self
-
-    def apply_function_to_position(self, function):
-        self.move_to(function(self.get_center()))
-        return self
-
-    def apply_function_to_submobject_positions(self, function):
-        for submob in self.submobjects:
-            submob.apply_function_to_position(function)
-        return self
-
-    def apply_matrix(self, matrix, **kwargs):
-        # Default to applying matrix about the origin, not mobjects center
-        if ("about_point" not in kwargs) and ("about_edge" not in kwargs):
-            kwargs["about_point"] = ORIGIN
-        full_matrix = np.identity(self.dim)
-        matrix = np.array(matrix)
-        full_matrix[: matrix.shape[0], : matrix.shape[1]] = matrix
-        self.apply_points_function_about_point(
-            lambda points: np.dot(points, full_matrix.T), **kwargs
-        )
-        return self
-
-    def apply_complex_function(self, function, **kwargs):
-        """Applies a complex function to a :class:`Mobject`.
-        The x and y coordinates correspond to the real and imaginary parts respectively.
-
-        Example
-        -------
-
-        .. manim:: ApplyFuncExample
-
-            class ApplyFuncExample(Scene):
-                def construct(self):
-                    circ = Circle().scale(1.5)
-                    circ_ref = circ.copy()
-                    circ.apply_complex_function(
-                        lambda x: np.exp(x*1j)
-                    )
-                    t = ValueTracker(0)
-                    circ.add_updater(
-                        lambda x: x.become(circ_ref.copy().apply_complex_function(
-                            lambda x: np.exp(x+t.get_value()*1j)
-                        )).set_color(BLUE)
-                    )
-                    self.add(circ_ref)
-                    self.play(TransformFromCopy(circ_ref, circ))
-                    self.play(t.animate.set_value(TAU), run_time=3)
-        """
-
-        def R3_func(point):
-            x, y, z = point
-            xy_complex = function(complex(x, y))
-            return [xy_complex.real, xy_complex.imag, z]
-
-        return self.apply_function(R3_func)
-
-    def wag(self, direction=RIGHT, axis=DOWN, wag_factor=1.0):
-        for mob in self.family_members_with_points():
-            alphas = np.dot(mob.points, np.transpose(axis))
-            alphas -= min(alphas)
-            alphas /= max(alphas)
-            alphas = alphas ** wag_factor
-            mob.points += np.dot(
-                alphas.reshape((len(alphas), 1)),
-                np.array(direction).reshape((1, mob.dim)),
-            )
-        return self
-
-    def reverse_points(self):
-        for mob in self.family_members_with_points():
-            mob.apply_over_attr_arrays(lambda arr: np.array(list(reversed(arr))))
-        return self
-
-    def repeat(self, count: int):
-        """This can make transition animations nicer"""
-
-        def repeat_array(array):
-            return reduce(lambda a1, a2: np.append(a1, a2, axis=0), [array] * count)
-
-        for mob in self.family_members_with_points():
-            mob.apply_over_attr_arrays(repeat_array)
-        return self
-
-    # In place operations.
-    # Note, much of these are now redundant with default behavior of
-    # above methods
-
-    def apply_points_function_about_point(
-        self,
-        func,
-        about_point=None,
-        about_edge=None,
-    ):
-        if about_point is None:
-            if about_edge is None:
-                about_edge = ORIGIN
-            about_point = self.get_critical_point(about_edge)
-        for mob in self.family_members_with_points():
-            mob.points -= about_point
-            mob.points = func(mob.points)
-            mob.points += about_point
-        return self
-
-    def pose_at_angle(self, **kwargs):
-        self.rotate(TAU / 14, RIGHT + UP, **kwargs)
-        return self
-
-    # Positioning methods
-
-    def center(self):
-        self.shift(-self.get_center())
-        return self
-
-    def align_on_border(self, direction, buff=DEFAULT_MOBJECT_TO_EDGE_BUFFER):
-        """Direction just needs to be a vector pointing towards side or
-        corner in the 2d plane.
-        """
-        target_point = np.sign(direction) * (
-            config["frame_x_radius"],
-            config["frame_y_radius"],
-            0,
-        )
-        point_to_align = self.get_critical_point(direction)
-        shift_val = target_point - point_to_align - buff * np.array(direction)
-        shift_val = shift_val * abs(np.sign(direction))
-        self.shift(shift_val)
-        return self
-
-    def to_corner(self, corner=LEFT + DOWN, buff=DEFAULT_MOBJECT_TO_EDGE_BUFFER):
-        return self.align_on_border(corner, buff)
-
-    def to_edge(self, edge=LEFT, buff=DEFAULT_MOBJECT_TO_EDGE_BUFFER):
-        return self.align_on_border(edge, buff)
-
-    def next_to(
-        self,
-        mobject_or_point,
-        direction=RIGHT,
-        buff=DEFAULT_MOBJECT_TO_MOBJECT_BUFFER,
-        aligned_edge=ORIGIN,
-        submobject_to_align=None,
-        index_of_submobject_to_align=None,
-        coor_mask=np.array([1, 1, 1]),
-    ):
-        """Move this :class:`~.Mobject` next to another's :class:`~.Mobject` or coordinate.
-
-        Examples
-        --------
-
-        .. manim:: GeometricShapes
-            :save_last_frame:
-
-            class GeometricShapes(Scene):
-                def construct(self):
-                    d = Dot()
-                    c = Circle()
-                    s = Square()
-                    t = Triangle()
-                    d.next_to(c, RIGHT)
-                    s.next_to(c, LEFT)
-                    t.next_to(c, DOWN)
-                    self.add(d, c, s, t)
-
-        """
-        if isinstance(mobject_or_point, Mobject):
-            mob = mobject_or_point
-            if index_of_submobject_to_align is not None:
-                target_aligner = mob[index_of_submobject_to_align]
-            else:
-                target_aligner = mob
-            target_point = target_aligner.get_critical_point(aligned_edge + direction)
-        else:
-            target_point = mobject_or_point
-        if submobject_to_align is not None:
-            aligner = submobject_to_align
-        elif index_of_submobject_to_align is not None:
-            aligner = self[index_of_submobject_to_align]
-        else:
-            aligner = self
-        point_to_align = aligner.get_critical_point(aligned_edge - direction)
-        self.shift((target_point - point_to_align + buff * direction) * coor_mask)
-        return self
-
-    def shift_onto_screen(self, **kwargs):
-        space_lengths = [config["frame_x_radius"], config["frame_y_radius"]]
-        for vect in UP, DOWN, LEFT, RIGHT:
-            dim = np.argmax(np.abs(vect))
-            buff = kwargs.get("buff", DEFAULT_MOBJECT_TO_EDGE_BUFFER)
-            max_val = space_lengths[dim] - buff
-            edge_center = self.get_edge_center(vect)
-            if np.dot(edge_center, vect) > max_val:
-                self.to_edge(vect, **kwargs)
-        return self
-
-    def is_off_screen(self):
-        if self.get_left()[0] > config["frame_x_radius"]:
-            return True
-        if self.get_right()[0] < -config["frame_x_radius"]:
-            return True
-        if self.get_bottom()[1] > config["frame_y_radius"]:
-            return True
-        if self.get_top()[1] < -config["frame_y_radius"]:
-            return True
-        return False
-
-    def stretch_about_point(self, factor, dim, point):
-        return self.stretch(factor, dim, about_point=point)
-
-    def rescale_to_fit(self, length, dim, stretch=False, **kwargs):
-        old_length = self.length_over_dim(dim)
-        if old_length == 0:
-            return self
-        if stretch:
-            self.stretch(length / old_length, dim, **kwargs)
-        else:
-            self.scale(length / old_length, **kwargs)
-        return self
-
-    def scale_to_fit_width(self, width, **kwargs):
-        """Scales the :class:`~.Mobject` to fit a width while keeping height/depth proportional.
-
-        Returns
-        -------
-        :class:`Mobject`
-            ``self``
-
-        Examples
-        --------
-        ::
-
-            >>> from manim import *
-            >>> sq = Square()
-            >>> sq.height
-            2.0
-            >>> sq.scale_to_fit_width(5)
-            Square
-            >>> sq.width
-            5.0
-            >>> sq.height
-            5.0
-        """
-
-        return self.rescale_to_fit(width, 0, stretch=False, **kwargs)
-
-    def stretch_to_fit_width(self, width, **kwargs):
-        """Stretches the :class:`~.Mobject` to fit a width, not keeping height/depth proportional.
-
-        Returns
-        -------
-        :class:`Mobject`
-            ``self``
-
-        Examples
-        --------
-        ::
-
-            >>> from manim import *
-            >>> sq = Square()
-            >>> sq.height
-            2.0
-            >>> sq.stretch_to_fit_width(5)
-            Square
-            >>> sq.width
-            5.0
-            >>> sq.height
-            2.0
-        """
-
-        return self.rescale_to_fit(width, 0, stretch=True, **kwargs)
-
-    def scale_to_fit_height(self, height, **kwargs):
-        """Scales the :class:`~.Mobject` to fit a height while keeping width/depth proportional.
-
-        Returns
-        -------
-        :class:`Mobject`
-            ``self``
-
-        Examples
-        --------
-        ::
-
-            >>> from manim import *
-            >>> sq = Square()
-            >>> sq.width
-            2.0
-            >>> sq.scale_to_fit_height(5)
-            Square
-            >>> sq.height
-            5.0
-            >>> sq.width
-            5.0
-        """
-
-        return self.rescale_to_fit(height, 1, stretch=False, **kwargs)
-
-    def stretch_to_fit_height(self, height, **kwargs):
-        """Stretches the :class:`~.Mobject` to fit a height, not keeping width/depth proportional.
-
-        Returns
-        -------
-        :class:`Mobject`
-            ``self``
-
-        Examples
-        --------
-        ::
-
-            >>> from manim import *
-            >>> sq = Square()
-            >>> sq.width
-            2.0
-            >>> sq.stretch_to_fit_height(5)
-            Square
-            >>> sq.height
-            5.0
-            >>> sq.width
-            2.0
-        """
-
-        return self.rescale_to_fit(height, 1, stretch=True, **kwargs)
-
-    def scale_to_fit_depth(self, depth, **kwargs):
-        """Scales the :class:`~.Mobject` to fit a depth while keeping width/height proportional."""
-
-        return self.rescale_to_fit(depth, 2, stretch=False, **kwargs)
-
-    def stretch_to_fit_depth(self, depth, **kwargs):
-        """Stretches the :class:`~.Mobject` to fit a depth, not keeping width/height proportional."""
-
-        return self.rescale_to_fit(depth, 2, stretch=True, **kwargs)
-
-    def set_coord(self, value, dim, direction=ORIGIN):
-        curr = self.get_coord(dim, direction)
-        shift_vect = np.zeros(self.dim)
-        shift_vect[dim] = value - curr
-        self.shift(shift_vect)
-        return self
-
-    def set_x(self, x, direction=ORIGIN):
-        """Set x value of the center of the :class:`~.Mobject` (``int`` or ``float``)"""
-        return self.set_coord(x, 0, direction)
-
-    def set_y(self, y, direction=ORIGIN):
-        """Set y value of the center of the :class:`~.Mobject` (``int`` or ``float``)"""
-        return self.set_coord(y, 1, direction)
-
-    def set_z(self, z, direction=ORIGIN):
-        """Set z value of the center of the :class:`~.Mobject` (``int`` or ``float``)"""
-        return self.set_coord(z, 2, direction)
-
-    def space_out_submobjects(self, factor=1.5, **kwargs):
-        self.scale(factor, **kwargs)
-        for submob in self.submobjects:
-            submob.scale(1.0 / factor)
-        return self
-
-    def move_to(
-        self,
-        point_or_mobject,
-        aligned_edge=ORIGIN,
-        coor_mask=np.array([1, 1, 1]),
-    ):
-        """Move center of the :class:`~.Mobject` to certain coordinate."""
-        if isinstance(point_or_mobject, Mobject):
-            target = point_or_mobject.get_critical_point(aligned_edge)
-        else:
-            target = point_or_mobject
-        point_to_align = self.get_critical_point(aligned_edge)
-        self.shift((target - point_to_align) * coor_mask)
-        return self
-
-    def replace(self, mobject, dim_to_match=0, stretch=False):
-        if not mobject.get_num_points() and not mobject.submobjects:
-            raise Warning("Attempting to replace mobject with no points")
-        if stretch:
-            self.stretch_to_fit_width(mobject.width)
-            self.stretch_to_fit_height(mobject.height)
-        else:
-            self.rescale_to_fit(
-                mobject.length_over_dim(dim_to_match),
-                dim_to_match,
-                stretch=False,
-            )
-        self.shift(mobject.get_center() - self.get_center())
-        return self
-
-    def surround(
-        self,
-        mobject: "Mobject",
-        dim_to_match=0,
-        stretch=False,
-        buff=MED_SMALL_BUFF,
-    ):
-        self.replace(mobject, dim_to_match, stretch)
-        length = mobject.length_over_dim(dim_to_match)
-        self.scale((length + buff) / length)
-        return self
-
-    def put_start_and_end_on(self, start, end):
-        curr_start, curr_end = self.get_start_and_end()
-        curr_vect = curr_end - curr_start
-        if np.all(curr_vect == 0):
-            raise Exception("Cannot position endpoints of closed loop")
-        target_vect = np.array(end) - np.array(start)
-        axis = (
-            normalize(np.cross(curr_vect, target_vect))
-            if np.linalg.norm(np.cross(curr_vect, target_vect)) != 0
-            else OUT
-        )
-        self.scale(
-            np.linalg.norm(target_vect) / np.linalg.norm(curr_vect),
-            about_point=curr_start,
-        )
-        self.rotate(
-            angle_between_vectors(curr_vect, target_vect),
-            about_point=curr_start,
-            axis=axis,
-        )
-        self.shift(start - curr_start)
-        return self
-
-    # Background rectangle
-    def add_background_rectangle(
-        self, color: Optional[Colors] = None, opacity: float = 0.75, **kwargs
-    ):
-        """Add a BackgroundRectangle as submobject.
-
-        The BackgroundRectangle is added behind other submobjects.
-
-        This can be used to increase the mobjects visibility in front of a noisy background.
-
-        Parameters
-        ----------
-        color
-            The color of the BackgroundRectangle
-        opacity
-            The opacity of the BackgroundRectangle
-        kwargs
-            Additional keyword arguments passed to the BackgroundRectangle constructor
-
-
-        Returns
-        -------
-        :class:`Mobject`
-            ``self``
-
-        See Also
-        --------
-        :meth:`add_to_back`
-        :class:`~.BackgroundRectangle`
-
-        """
-
-        # TODO, this does not behave well when the mobject has points,
-        # since it gets displayed on top
-        from ..mobject.shape_matchers import BackgroundRectangle
-
-        self.background_rectangle = BackgroundRectangle(
-            self, color=color, fill_opacity=opacity, **kwargs
-        )
-        self.add_to_back(self.background_rectangle)
-        return self
-
-    def add_background_rectangle_to_submobjects(self, **kwargs):
-        for submobject in self.submobjects:
-            submobject.add_background_rectangle(**kwargs)
-        return self
-
-    def add_background_rectangle_to_family_members_with_points(self, **kwargs):
-        for mob in self.family_members_with_points():
-            mob.add_background_rectangle(**kwargs)
-        return self
-
-    # Color functions
-
-    def set_color(self, color: Color = YELLOW_C, family: bool = True):
-        """Condition is function which takes in one arguments, (x, y, z).
-        Here it just recurses to submobjects, but in subclasses this
-        should be further implemented based on the the inner workings
-        of color
-        """
-        if family:
-            for submob in self.submobjects:
-                submob.set_color(color, family=family)
-        self.color = Color(color)
-        return self
-
-    def set_color_by_gradient(self, *colors):
-        self.set_submobject_colors_by_gradient(*colors)
-        return self
-
-    def set_colors_by_radial_gradient(
-        self,
-        center=None,
-        radius=1,
-        inner_color=WHITE,
-        outer_color=BLACK,
-    ):
-        self.set_submobject_colors_by_radial_gradient(
-            center,
-            radius,
-            inner_color,
-            outer_color,
-        )
-        return self
-
-    def set_submobject_colors_by_gradient(self, *colors):
-        if len(colors) == 0:
-            raise ValueError("Need at least one color")
-        elif len(colors) == 1:
-            return self.set_color(*colors)
-
-        mobs = self.family_members_with_points()
-        new_colors = color_gradient(colors, len(mobs))
-
-        for mob, color in zip(mobs, new_colors):
-            mob.set_color(color, family=False)
-        return self
-
-    def set_submobject_colors_by_radial_gradient(
-        self,
-        center=None,
-        radius=1,
-        inner_color=WHITE,
-        outer_color=BLACK,
-    ):
-        if center is None:
-            center = self.get_center()
-
-        for mob in self.family_members_with_points():
-            t = np.linalg.norm(mob.get_center() - center) / radius
-            t = min(t, 1)
-            mob_color = interpolate_color(inner_color, outer_color, t)
-            mob.set_color(mob_color, family=False)
-
-        return self
-
-    def to_original_color(self):
-        self.set_color(self.color)
-        return self
-
-    def fade_to(self, color, alpha, family=True):
-        if self.get_num_points() > 0:
-            new_color = interpolate_color(self.get_color(), color, alpha)
-            self.set_color(new_color, family=False)
-        if family:
-            for submob in self.submobjects:
-                submob.fade_to(color, alpha)
-        return self
-
-    def fade(self, darkness=0.5, family=True):
-        if family:
-            for submob in self.submobjects:
-                submob.fade(darkness, family)
-        return self
-
-    def get_color(self):
-        """Returns the color of the :class:`~.Mobject`"""
-        return self.color
-
-    ##
-
-    def save_state(self):
-        """Save the current state (position, color & size). Can be restored with :meth:`~.Mobject.restore`."""
-        if hasattr(self, "saved_state"):
-            # Prevent exponential growth of data
-            self.saved_state = None
-        self.saved_state = self.copy()
-
-        return self
-
-    def restore(self):
-        """Restores the state that was previously saved with :meth:`~.Mobject.save_state`."""
-        if not hasattr(self, "saved_state") or self.save_state is None:
-            raise Exception("Trying to restore without having saved")
-        self.become(self.saved_state)
-        return self
-
-    ##
-
-    def reduce_across_dimension(self, points_func, reduce_func, dim):
-        points = self.get_all_points()
-        if points is None or len(points) == 0:
-            # Note, this default means things like empty VGroups
-            # will appear to have a center at [0, 0, 0]
-            return 0
-        values = points_func(points[:, dim])
-        return reduce_func(values)
-
-    def nonempty_submobjects(self):
-        return [
-            submob
-            for submob in self.submobjects
-            if len(submob.submobjects) != 0 or len(submob.points) != 0
-        ]
-
-    def get_merged_array(self, array_attr):
-        result = getattr(self, array_attr)
-        for submob in self.submobjects:
-            result = np.append(result, submob.get_merged_array(array_attr), axis=0)
-        return result
-
-    def get_all_points(self):
-        return self.get_merged_array("points")
-
-    # Getters
-
-    def get_points_defining_boundary(self):
-        return self.get_all_points()
-
-    def get_num_points(self):
-        return len(self.points)
-
-    def get_extremum_along_dim(self, points=None, dim=0, key=0):
-        if points is None:
-            points = self.get_points_defining_boundary()
-        values = points[:, dim]
-        if key < 0:
-            return np.min(values)
-        elif key == 0:
-            return (np.min(values) + np.max(values)) / 2
-        else:
-            return np.max(values)
-
-    def get_critical_point(self, direction):
-        """Picture a box bounding the :class:`~.Mobject`.  Such a box has
-        9 'critical points': 4 corners, 4 edge center, the
-        center. This returns one of them, along the given direction.
-
-        ::
-
-            sample = Arc(start_angle=PI/7, angle = PI/5)
-
-            # These are all equivalent
-            max_y_1 = sample.get_top()[1]
-            max_y_2 = sample.get_critical_point(UP)[1]
-            max_y_3 = sample.get_extremum_along_dim(dim=1, key=1)
-
-        """
-        result = np.zeros(self.dim)
-        all_points = self.get_points_defining_boundary()
-        if len(all_points) == 0:
-            return result
-        for dim in range(self.dim):
-            result[dim] = self.get_extremum_along_dim(
-                all_points,
-                dim=dim,
-                key=direction[dim],
-            )
-        return result
-
-    # Pseudonyms for more general get_critical_point method
-
-    def get_edge_center(self, direction) -> np.ndarray:
-        """Get edge coordinates for certain direction."""
-        return self.get_critical_point(direction)
-
-    def get_corner(self, direction) -> np.ndarray:
-        """Get corner coordinates for certain direction."""
-        return self.get_critical_point(direction)
-
-    def get_center(self) -> np.ndarray:
-        """Get center coordinates"""
-        return self.get_critical_point(np.zeros(self.dim))
-
-    def get_center_of_mass(self):
-        return np.apply_along_axis(np.mean, 0, self.get_all_points())
-
-    def get_boundary_point(self, direction):
-        all_points = self.get_points_defining_boundary()
-        index = np.argmax(np.dot(all_points, np.array(direction).T))
-        return all_points[index]
-
-    def get_midpoint(self) -> np.ndarray:
-        """Get coordinates of the middle of the path that forms the  :class:`~.Mobject`.
-
-        Examples
-        --------
-
-        .. manim:: AngleMidPoint
-            :save_last_frame:
-
-            class AngleMidPoint(Scene):
-                def construct(self):
-                    line1 = Line(ORIGIN, 2*RIGHT)
-                    line2 = Line(ORIGIN, 2*RIGHT).rotate_about_origin(80*DEGREES)
-
-                    a = Angle(line1, line2, radius=1.5, other_angle=False)
-                    d = Dot(a.get_midpoint()).set_color(RED)
-
-                    self.add(line1, line2, a, d)
-                    self.wait()
-
-        """
-        return self.point_from_proportion(0.5)
-
-    def get_top(self) -> np.ndarray:
-        """Get top coordinates of a box bounding the :class:`~.Mobject`"""
-        return self.get_edge_center(UP)
-
-    def get_bottom(self) -> np.ndarray:
-        """Get bottom coordinates of a box bounding the :class:`~.Mobject`"""
-        return self.get_edge_center(DOWN)
-
-    def get_right(self) -> np.ndarray:
-        """Get right coordinates of a box bounding the :class:`~.Mobject`"""
-        return self.get_edge_center(RIGHT)
-
-    def get_left(self) -> np.ndarray:
-        """Get left coordinates of a box bounding the :class:`~.Mobject`"""
-        return self.get_edge_center(LEFT)
-
-    def get_zenith(self) -> np.ndarray:
-        """Get zenith coordinates of a box bounding a 3D :class:`~.Mobject`."""
-        return self.get_edge_center(OUT)
-
-    def get_nadir(self) -> np.ndarray:
-        """Get nadir (opposite the zenith) coordinates of a box bounding a 3D :class:`~.Mobject`."""
-        return self.get_edge_center(IN)
-
-    def length_over_dim(self, dim):
-        """Measure the length of an :class:`~.Mobject` in a certain direction."""
-        return (
-            self.reduce_across_dimension(
-                np.max,
-                np.max,
-                dim,
-            )
-            - self.reduce_across_dimension(np.min, np.min, dim)
-        )
-
-    def get_coord(self, dim, direction=ORIGIN):
-        """Meant to generalize ``get_x``, ``get_y`` and ``get_z``"""
-        return self.get_extremum_along_dim(dim=dim, key=direction[dim])
-
-    def get_x(self, direction=ORIGIN) -> np.float64:
-        """Returns x coordinate of the center of the :class:`~.Mobject` as ``float``"""
-        return self.get_coord(0, direction)
-
-    def get_y(self, direction=ORIGIN) -> np.float64:
-        """Returns y coordinate of the center of the :class:`~.Mobject` as ``float``"""
-        return self.get_coord(1, direction)
-
-    def get_z(self, direction=ORIGIN) -> np.float64:
-        """Returns z coordinate of the center of the :class:`~.Mobject` as ``float``"""
-        return self.get_coord(2, direction)
-
-    def get_start(self):
-        """Returns the point, where the stroke that surrounds the :class:`~.Mobject` starts."""
-        self.throw_error_if_no_points()
-        return np.array(self.points[0])
-
-    def get_end(self):
-        """Returns the point, where the stroke that surrounds the :class:`~.Mobject` ends."""
-        self.throw_error_if_no_points()
-        return np.array(self.points[-1])
-
-    def get_start_and_end(self):
-        """Returns starting and ending point of a stroke as a ``tuple``."""
-        return self.get_start(), self.get_end()
-
-    def point_from_proportion(self, alpha):
-        raise NotImplementedError("Please override in a child class.")
-
-    def proportion_from_point(self, point):
-        raise NotImplementedError("Please override in a child class.")
-
-    def get_pieces(self, n_pieces):
-        template = self.copy()
-        template.submobjects = []
-        alphas = np.linspace(0, 1, n_pieces + 1)
-        return Group(
-            *(
-                template.copy().pointwise_become_partial(self, a1, a2)
-                for a1, a2 in zip(alphas[:-1], alphas[1:])
-            )
-        )
-
-    def get_z_index_reference_point(self):
-        # TODO, better place to define default z_index_group?
-        z_index_group = getattr(self, "z_index_group", self)
-        return z_index_group.get_center()
-
-    def has_points(self) -> bool:
-        """Check if :class:`~.Mobject` contains points."""
-        return len(self.points) > 0
-
-    def has_no_points(self) -> bool:
-        """Check if :class:`~.Mobject` *does not* contains points."""
-        return not self.has_points()
-
-    # Match other mobject properties
-
-    def match_color(self, mobject: "Mobject"):
-        """Match the color with the color of another :class:`~.Mobject`."""
-        return self.set_color(mobject.get_color())
-
-    def match_dim_size(self, mobject: "Mobject", dim, **kwargs):
-        """Match the specified dimension with the dimension of another :class:`~.Mobject`."""
-        return self.rescale_to_fit(mobject.length_over_dim(dim), dim, **kwargs)
-
-    def match_width(self, mobject: "Mobject", **kwargs):
-        """Match the width with the width of another :class:`~.Mobject`."""
-        return self.match_dim_size(mobject, 0, **kwargs)
-
-    def match_height(self, mobject: "Mobject", **kwargs):
-        """Match the height with the height of another :class:`~.Mobject`."""
-        return self.match_dim_size(mobject, 1, **kwargs)
-
-    def match_depth(self, mobject: "Mobject", **kwargs):
-        """Match the depth with the depth of another :class:`~.Mobject`."""
-        return self.match_dim_size(mobject, 2, **kwargs)
-
-    def match_coord(self, mobject: "Mobject", dim, direction=ORIGIN):
-        """Match the coordinates with the coordinates of another :class:`~.Mobject`."""
-        return self.set_coord(
-            mobject.get_coord(dim, direction),
-            dim=dim,
-            direction=direction,
-        )
-
-    def match_x(self, mobject: "Mobject", direction=ORIGIN):
-        """Match x coord. to the x coord. of another :class:`~.Mobject`."""
-        return self.match_coord(mobject, 0, direction)
-
-    def match_y(self, mobject: "Mobject", direction=ORIGIN):
-        """Match y coord. to the x coord. of another :class:`~.Mobject`."""
-        return self.match_coord(mobject, 1, direction)
-
-    def match_z(self, mobject: "Mobject", direction=ORIGIN):
-        """Match z coord. to the x coord. of another :class:`~.Mobject`."""
-        return self.match_coord(mobject, 2, direction)
-
-    def align_to(
-        self,
-        mobject_or_point: Union["Mobject", np.ndarray, List],
-        direction=ORIGIN,
-        alignment_vect=UP,
-    ):
-        """Aligns mobject to another :class:`~.Mobject` in a certain direction.
-
-        Examples:
-        mob1.align_to(mob2, UP) moves mob1 vertically so that its
-        top edge lines ups with mob2's top edge.
-
-        mob1.align_to(mob2, alignment_vect = RIGHT) moves mob1
-        horizontally so that it's center is directly above/below
-        the center of mob2
-        """
-        if isinstance(mobject_or_point, Mobject):
-            point = mobject_or_point.get_critical_point(direction)
-        else:
-            point = mobject_or_point
-
-        for dim in range(self.dim):
-            if direction[dim] != 0:
-                self.set_coord(point[dim], dim, direction)
-        return self
-
-    # Family matters
-
-    def __getitem__(self, value):
-        self_list = self.split()
-        if isinstance(value, slice):
-            GroupClass = self.get_group_class()
-            return GroupClass(*self_list.__getitem__(value))
-        return self_list.__getitem__(value)
-
-    def __iter__(self):
-        return iter(self.split())
-
-    def __len__(self):
-        return len(self.split())
-
-    def get_group_class(self):
-        return Group
-
-    def split(self):
-        result = [self] if len(self.points) > 0 else []
-        return result + self.submobjects
-
-    def get_family(self, recurse=True):
-        sub_families = list(map(Mobject.get_family, self.submobjects))
-        all_mobjects = [self] + list(it.chain(*sub_families))
-        return remove_list_redundancies(all_mobjects)
-
-    def family_members_with_points(self):
-        return [m for m in self.get_family() if m.get_num_points() > 0]
-
-    def arrange(
-        self,
-        direction: Sequence[float] = RIGHT,
-        buff=DEFAULT_MOBJECT_TO_MOBJECT_BUFFER,
-        center=True,
-        **kwargs,
-    ):
+    def arrange(self, direction=RIGHT, center=True, **kwargs):
         """Sorts :class:`~.Mobject` next to each other on screen.
 
         Examples
@@ -2131,11 +865,11 @@ class Mobject:
                     s2 = Square()
                     s3 = Square()
                     s4 = Square()
-                    x = VGroup(s1, s2, s3, s4).set_x(0).arrange(buff=1.0)
+                    x = OpenGLVGroup(s1, s2, s3, s4).set_x(0).arrange(buff=1.0)
                     self.add(x)
         """
         for m1, m2 in zip(self.submobjects, self.submobjects[1:]):
-            m2.next_to(m1, direction, buff, **kwargs)
+            m2.next_to(m1, direction, **kwargs)
         if center:
             self.center()
         return self
@@ -2152,7 +886,7 @@ class Mobject:
         col_widths: Optional[Iterable[Optional[float]]] = None,
         flow_order: str = "rd",
         **kwargs,
-    ):
+    ) -> "Mobject":
         """Arrange submobjects in a grid.
 
         Parameters
@@ -2185,8 +919,19 @@ class Mobject:
 
         Returns
         -------
-        :class:`Mobject`
-            ``self``
+        Mobject
+            The mobject.
+
+        NOTES
+        -----
+
+        If only one of ``cols`` and ``rows`` is set implicitly, the other one will be chosen big
+        enough to fit all submobjects. If neither is set, they will be chosen to be about the same,
+        tending towards ``cols`` > ``rows`` (simply because videos are wider than they are high).
+
+        If both ``cell_alignment`` and ``row_alignments`` / ``col_alignments`` are
+        defined, the latter has higher priority.
+
 
         Raises
         ------
@@ -2195,15 +940,6 @@ class Mobject:
         ValueError
             If :code:`cols`, :code:`col_alignments` and :code:`col_widths` or :code:`rows`,
             :code:`row_alignments` and :code:`row_heights` have mismatching sizes.
-
-        Notes
-        -----
-        If only one of ``cols`` and ``rows`` is set implicitly, the other one will be chosen big
-        enough to fit all submobjects. If neither is set, they will be chosen to be about the same,
-        tending towards ``cols`` > ``rows`` (simply because videos are wider than they are high).
-
-        If both ``cell_alignment`` and ``row_alignments`` / ``col_alignments`` are
-        defined, the latter has higher priority.
 
         Examples
         --------
@@ -2382,22 +1118,51 @@ class Mobject:
         self.move_to(start_pos)
         return self
 
+    def get_grid(self, n_rows, n_cols, height=None, **kwargs):
+        """
+        Returns a new mobject containing multiple copies of this one
+        arranged in a grid
+        """
+        grid = self.duplicate(n_rows * n_cols)
+        grid.arrange_in_grid(n_rows, n_cols, **kwargs)
+        if height is not None:
+            grid.set_height(height)
+        return grid
+
+    def duplicate(self, n: int):
+        """Returns an :class:`~.OpenGLVGroup` containing ``n`` copies of the mobject."""
+        return self.get_group_class()(*[self.copy() for _ in range(n)])
+
     def sort(self, point_to_num_func=lambda p: p[0], submob_func=None):
         """Sorts the list of :attr:`submobjects` by a function defined by ``submob_func``."""
-        if submob_func is None:
-
-            def submob_func(m):
-                return point_to_num_func(m.get_center())
-
-        self.submobjects.sort(key=submob_func)
+        if submob_func is not None:
+            self.submobjects.sort(key=submob_func)
+        else:
+            self.submobjects.sort(key=lambda m: point_to_num_func(m.get_center()))
         return self
 
-    def shuffle(self, recursive=False):
-        """Shuffles the list of :attr:`submobjects`."""
-        if recursive:
+    def shuffle(self, recurse=False):
+        """Shuffles the order of :attr:`submobjects`
+
+        Examples
+        --------
+
+        .. manim:: ShuffleSubmobjectsExample
+
+            class ShuffleSubmobjectsExample(Scene):
+                def construct(self):
+                    s= OpenGLVGroup(*[Dot().shift(i*0.1*RIGHT) for i in range(-20,20)])
+                    s2= s.copy()
+                    s2.shuffle()
+                    s2.shift(DOWN)
+                    self.play(Write(s), Write(s2))
+        """
+        if recurse:
             for submob in self.submobjects:
-                submob.shuffle(recursive=True)
+                submob.shuffle(recurse=True)
         random.shuffle(self.submobjects)
+        self.assemble_family()
+        return self
 
     def invert(self, recursive=False):
         """Inverts the list of :attr:`submobjects`.
@@ -2425,135 +1190,1121 @@ class Mobject:
                 submob.invert(recursive=True)
         list.reverse(self.submobjects)
 
-    # Just here to keep from breaking old scenes.
-    def arrange_submobjects(self, *args, **kwargs):
-        """Arrange the position of :attr:`submobjects` with a small buffer.
+    # Copying
+
+    def copy(self, shallow: bool = False):
+        """Create and return an identical copy of the :class:`Mobject` including all
+        :attr:`submobjects`.
+
+        Returns
+        -------
+        :class:`Mobject`
+            The copy.
+
+        Parameters
+        ----------
+        shallow
+            Controls whether a shallow copy is returned.
+
+        Note
+        ----
+        The clone is initially not visible in the Scene, even if the original was.
+        """
+        if not shallow:
+            return self.deepcopy()
+
+        # TODO, either justify reason for shallow copy, or
+        # remove this redundancy everywhere
+        # return self.deepcopy()
+
+        parents = self.parents
+        self.parents = []
+        copy_mobject = copy.copy(self)
+        self.parents = parents
+
+        copy_mobject.data = dict(self.data)
+        for key in self.data:
+            copy_mobject.data[key] = self.data[key].copy()
+
+        # TODO, are uniforms ever numpy arrays?
+        copy_mobject.uniforms = dict(self.uniforms)
+
+        copy_mobject.submobjects = []
+        copy_mobject.add(*(sm.copy() for sm in self.submobjects))
+        copy_mobject.match_updaters(self)
+
+        copy_mobject.needs_new_bounding_box = self.needs_new_bounding_box
+
+        # Make sure any mobject or numpy array attributes are copied
+        family = self.get_family()
+        for attr, value in list(self.__dict__.items()):
+            if (
+                isinstance(value, Mobject)
+                and value in family
+                and value is not self
+            ):
+                setattr(copy_mobject, attr, value.copy())
+            if isinstance(value, np.ndarray):
+                setattr(copy_mobject, attr, value.copy())
+            # if isinstance(value, ShaderWrapper):
+            #     setattr(copy_mobject, attr, value.copy())
+        return copy_mobject
+
+    def deepcopy(self):
+        parents = self.parents
+        self.parents = []
+        result = copy.deepcopy(self)
+        self.parents = parents
+        return result
+
+    def generate_target(self, use_deepcopy: bool = False):
+        self.target = None  # Prevent exponential explosion
+        if use_deepcopy:
+            self.target = self.deepcopy()
+        else:
+            self.target = self.copy()
+        return self.target
+
+    def save_state(self, use_deepcopy: bool = False):
+        """Save the current state (position, color & size). Can be restored with :meth:`~.Mobject.restore`."""
+        if hasattr(self, "saved_state"):
+            # Prevent exponential growth of data
+            self.saved_state = None
+        if use_deepcopy:
+            self.saved_state = self.deepcopy()
+        else:
+            self.saved_state = self.copy()
+        return self
+
+    def restore(self):
+        """Restores the state that was previously saved with :meth:`~.Mobject.save_state`."""
+        if not hasattr(self, "saved_state") or self.save_state is None:
+            raise Exception("Trying to restore without having saved")
+        self.become(self.saved_state)
+        return self
+
+    # Updating
+
+    def init_updaters(self):
+        self.time_based_updaters = []
+        self.non_time_updaters = []
+        self.has_updaters = False
+        self.updating_suspended = False
+
+    def update(self, dt=0, recurse=True):
+        if not self.has_updaters or self.updating_suspended:
+            return self
+        for updater in self.time_based_updaters:
+            updater(self, dt)
+        for updater in self.non_time_updaters:
+            updater(self)
+        if recurse:
+            for submob in self.submobjects:
+                submob.update(dt, recurse)
+        return self
+
+    def get_time_based_updaters(self):
+        return self.time_based_updaters
+
+    def has_time_based_updater(self):
+        return len(self.time_based_updaters) > 0
+
+    def get_updaters(self):
+        return self.time_based_updaters + self.non_time_updaters
+
+    def get_family_updaters(self):
+        return list(it.chain(*(sm.get_updaters() for sm in self.get_family())))
+
+    def add_updater(self, update_function, index=None, call_updater=True):
+        if "dt" in get_parameters(update_function):
+            updater_list = self.time_based_updaters
+        else:
+            updater_list = self.non_time_updaters
+
+        if index is None:
+            updater_list.append(update_function)
+        else:
+            updater_list.insert(index, update_function)
+
+        self.refresh_has_updater_status()
+        if call_updater:
+            self.update()
+        return self
+
+    def remove_updater(self, update_function):
+        for updater_list in [self.time_based_updaters, self.non_time_updaters]:
+            while update_function in updater_list:
+                updater_list.remove(update_function)
+        self.refresh_has_updater_status()
+        return self
+
+    def clear_updaters(self, recurse=True):
+        self.time_based_updaters = []
+        self.non_time_updaters = []
+        self.refresh_has_updater_status()
+        if recurse:
+            for submob in self.submobjects:
+                submob.clear_updaters()
+        return self
+
+    def match_updaters(self, mobject):
+        self.clear_updaters()
+        for updater in mobject.get_updaters():
+            self.add_updater(updater)
+        return self
+
+    def suspend_updating(self, recurse=True):
+        self.updating_suspended = True
+        if recurse:
+            for submob in self.submobjects:
+                submob.suspend_updating(recurse)
+        return self
+
+    def resume_updating(self, recurse=True, call_updater=True):
+        self.updating_suspended = False
+        if recurse:
+            for submob in self.submobjects:
+                submob.resume_updating(recurse)
+        for parent in self.parents:
+            parent.resume_updating(recurse=False, call_updater=False)
+        if call_updater:
+            self.update(dt=0, recurse=recurse)
+        return self
+
+    def refresh_has_updater_status(self):
+        self.has_updaters = any(mob.get_updaters() for mob in self.get_family())
+        return self
+
+    # Transforming operations
+
+    def shift(self, vector):
+        self.apply_points_function(
+            lambda points: points + vector,
+            about_edge=None,
+            works_on_bounding_box=True,
+        )
+        return self
+
+    def scale(
+        self,
+        scale_factor: float,
+        about_point: Optional[Sequence[float]] = None,
+        about_edge: Sequence[float] = ORIGIN,
+        **kwargs,
+    ) -> "Mobject":
+        r"""Scale the size by a factor.
+
+        Default behavior is to scale about the center of the mobject.
+        The argument about_edge can be a vector, indicating which side of
+        the mobject to scale about, e.g., mob.scale(about_edge = RIGHT)
+        scales about mob.get_right().
+
+        Otherwise, if about_point is given a value, scaling is done with
+        respect to that point.
+
+        Parameters
+        ----------
+        scale_factor
+            The scaling factor :math:`\alpha`. If :math:`0 < |\alpha| < 1`, the mobject
+            will shrink, and for :math:`|\alpha| > 1` it will grow. Furthermore,
+            if :math:`\alpha < 0`, the mobject is also flipped.
+        kwargs
+            Additional keyword arguments passed to
+            :meth:`apply_points_function_about_point`.
+
+        Returns
+        -------
+        Mobject
+            The scaled mobject.
 
         Examples
         --------
 
-        .. manim:: ArrangeSumobjectsExample
+        .. manim:: MobjectScaleExample
             :save_last_frame:
 
-            class ArrangeSumobjectsExample(Scene):
+            class MobjectScaleExample(Scene):
                 def construct(self):
-                    s= VGroup(*[Dot().shift(i*0.1*RIGHT*np.random.uniform(-1,1)+UP*np.random.uniform(-1,1)) for i in range(0,15)])
-                    s.shift(UP).set_color(BLUE)
-                    s2= s.copy().set_color(RED)
-                    s2.arrange_submobjects()
-                    s2.shift(DOWN)
-                    self.add(s,s2)
+                    f1 = Text("F")
+                    f2 = Text("F").scale(2)
+                    f3 = Text("F").scale(0.5)
+                    f4 = Text("F").scale(-1)
+
+                    vgroup = VGroup(f1, f2, f3, f4).arrange(6 * RIGHT)
+                    self.add(vgroup)
+
+        See also
+        --------
+        :meth:`move_to`
 
         """
-        return self.arrange(*args, **kwargs)
+        self.apply_points_function(
+            lambda points: scale_factor * points,
+            about_point=about_point,
+            about_edge=about_edge,
+            works_on_bounding_box=True,
+            **kwargs,
+        )
+        return self
 
-    def sort_submobjects(self, *args, **kwargs):
-        """Sort the :attr:`submobjects`"""
-        return self.sort(*args, **kwargs)
+    def stretch(self, factor, dim, **kwargs):
+        def func(points):
+            points[:, dim] *= factor
+            return points
 
-    def shuffle_submobjects(self, *args, **kwargs):
-        """Shuffles the order of :attr:`submobjects`
+        self.apply_points_function(func, works_on_bounding_box=True, **kwargs)
+        return self
+
+    def rotate_about_origin(self, angle, axis=OUT):
+        return self.rotate(angle, axis, about_point=ORIGIN)
+
+    def rotate(
+        self,
+        angle,
+        axis=OUT,
+        about_point: Optional[Sequence[float]] = None,
+        **kwargs,
+    ):
+        """Rotates the :class:`~.Mobject` about a certain point."""
+        rot_matrix_T = rotation_matrix_transpose(angle, axis)
+        self.apply_points_function(
+            lambda points: np.dot(points, rot_matrix_T),
+            about_point=about_point,
+            **kwargs,
+        )
+        return self
+
+    def flip(self, axis=UP, **kwargs):
+        """Flips/Mirrors an mobject about its center.
 
         Examples
         --------
 
-        .. manim:: ShuffleSubmobjectsExample
+        .. manim:: FlipExample
+            :save_last_frame:
 
-            class ShuffleSubmobjectsExample(Scene):
+            class FlipExample(Scene):
                 def construct(self):
-                    s= VGroup(*[Dot().shift(i*0.1*RIGHT) for i in range(-20,20)])
-                    s2= s.copy()
-                    s2.shuffle_submobjects()
-                    s2.shift(DOWN)
-                    self.play(Write(s), Write(s2))
+                    s= Line(LEFT, RIGHT+UP).shift(4*LEFT)
+                    self.add(s)
+                    s2= s.copy().flip()
+                    self.add(s2)
+
         """
-        return self.shuffle(*args, **kwargs)
+        return self.rotate(TAU / 2, axis, **kwargs)
 
-    # Alignment
-    def align_data(self, mobject: "Mobject"):
-        self.null_point_align(mobject)
-        self.align_submobjects(mobject)
-        self.align_points(mobject)
-        # Recurse
-        for m1, m2 in zip(self.submobjects, mobject.submobjects):
-            m1.align_data(m2)
-
-    def get_point_mobject(self, center=None):
-        """The simplest :class:`~.Mobject` to be transformed to or from self.
-        Should by a point of the appropriate type
-        """
-        msg = f"get_point_mobject not implemented for {self.__class__.__name__}"
-        raise NotImplementedError(msg)
-
-    def align_points(self, mobject):
-        count1 = self.get_num_points()
-        count2 = mobject.get_num_points()
-        if count1 < count2:
-            self.align_points_with_larger(mobject)
-        elif count2 < count1:
-            mobject.align_points_with_larger(self)
+    def apply_function(self, function, **kwargs):
+        # Default to applying matrix about the origin, not mobjects center
+        if len(kwargs) == 0:
+            kwargs["about_point"] = ORIGIN
+        self.apply_points_function(
+            lambda points: np.array([function(p) for p in points]), **kwargs
+        )
         return self
 
-    def align_points_with_larger(self, larger_mobject):
-        raise NotImplementedError("Please override in a child class.")
-
-    def align_submobjects(self, mobject):
-        mob1 = self
-        mob2 = mobject
-        n1 = len(mob1.submobjects)
-        n2 = len(mob2.submobjects)
-        mob1.add_n_more_submobjects(max(0, n2 - n1))
-        mob2.add_n_more_submobjects(max(0, n1 - n2))
+    def apply_function_to_position(self, function):
+        self.move_to(function(self.get_center()))
         return self
 
-    def null_point_align(self, mobject: "Mobject"):
-        """If a :class:`~.Mobject` with points is being aligned to
-        one without, treat both as groups, and push
-        the one with points into its own submobjects
-        list.
+    def apply_function_to_submobject_positions(self, function):
+        for submob in self.submobjects:
+            submob.apply_function_to_position(function)
+        return self
+
+    def apply_matrix(self, matrix, **kwargs):
+        # Default to applying matrix about the origin, not mobjects center
+        if ("about_point" not in kwargs) and ("about_edge" not in kwargs):
+            kwargs["about_point"] = ORIGIN
+        full_matrix = np.identity(self.dim)
+        matrix = np.array(matrix)
+        full_matrix[: matrix.shape[0], : matrix.shape[1]] = matrix
+        self.apply_points_function(
+            lambda points: np.dot(points, full_matrix.T), **kwargs
+        )
+        return self
+
+    def apply_complex_function(self, function, **kwargs):
+        """Applies a complex function to a :class:`Mobject`.
+        The x and y coordinates correspond to the real and imaginary parts respectively.
+
+        Example
+        -------
+
+        .. manim:: ApplyFuncExample
+
+            class ApplyFuncExample(Scene):
+                def construct(self):
+                    circ = Circle().scale(1.5)
+                    circ_ref = circ.copy()
+                    circ.apply_complex_function(
+                        lambda x: np.exp(x*1j)
+                    )
+                    t = ValueTracker(0)
+                    circ.add_updater(
+                        lambda x: x.become(circ_ref.copy().apply_complex_function(
+                            lambda x: np.exp(x+t.get_value()*1j)
+                        )).set_color(BLUE)
+                    )
+                    self.add(circ_ref)
+                    self.play(TransformFromCopy(circ_ref, circ))
+                    self.play(t.animate.set_value(TAU), run_time=3)
+        """
+
+        def R3_func(point):
+            x, y, z = point
+            xy_complex = function(complex(x, y))
+            return [xy_complex.real, xy_complex.imag, z]
+
+        return self.apply_function(R3_func)
+
+    def hierarchical_model_matrix(self):
+        if self.parent is None:
+            return self.model_matrix
+
+        model_matrices = [self.model_matrix]
+        current_object = self
+        while current_object.parent is not None:
+            model_matrices.append(current_object.parent.model_matrix)
+            current_object = current_object.parent
+        return np.linalg.multi_dot(list(reversed(model_matrices)))
+
+    def wag(self, direction=RIGHT, axis=DOWN, wag_factor=1.0):
+        for mob in self.family_members_with_points():
+            alphas = np.dot(mob.points, np.transpose(axis))
+            alphas -= min(alphas)
+            alphas /= max(alphas)
+            alphas = alphas ** wag_factor
+            mob.set_points(
+                mob.points
+                + np.dot(
+                    alphas.reshape((len(alphas), 1)),
+                    np.array(direction).reshape((1, mob.dim)),
+                ),
+            )
+        return self
+
+    # Positioning methods
+
+    def center(self):
+        """Moves the mobject to the center of the Scene."""
+        self.shift(-self.get_center())
+        return self
+
+    def align_on_border(self, direction, buff=DEFAULT_MOBJECT_TO_EDGE_BUFFER):
+        """
+        Direction just needs to be a vector pointing towards side or
+        corner in the 2d plane.
+        """
+        target_point = np.sign(direction) * (
+            config["frame_x_radius"],
+            config["frame_y_radius"],
+            0,
+        )
+        point_to_align = self.get_bounding_box_point(direction)
+        shift_val = target_point - point_to_align - buff * np.array(direction)
+        shift_val = shift_val * abs(np.sign(direction))
+        self.shift(shift_val)
+        return self
+
+    def to_corner(self, corner=LEFT + DOWN, buff=DEFAULT_MOBJECT_TO_EDGE_BUFFER):
+        return self.align_on_border(corner, buff)
+
+    def to_edge(self, edge=LEFT, buff=DEFAULT_MOBJECT_TO_EDGE_BUFFER):
+        return self.align_on_border(edge, buff)
+
+    def next_to(
+        self,
+        mobject_or_point,
+        direction=RIGHT,
+        buff=DEFAULT_MOBJECT_TO_MOBJECT_BUFFER,
+        aligned_edge=ORIGIN,
+        submobject_to_align=None,
+        index_of_submobject_to_align=None,
+        coor_mask=np.array([1, 1, 1]),
+    ):
+        """Move this :class:`~.Mobject` next to another's :class:`~.Mobject` or coordinate.
+
+        Examples
+        --------
+
+        .. manim:: GeometricShapes
+            :save_last_frame:
+
+            class GeometricShapes(Scene):
+                def construct(self):
+                    d = Dot()
+                    c = Circle()
+                    s = Square()
+                    t = Triangle()
+                    d.next_to(c, RIGHT)
+                    s.next_to(c, LEFT)
+                    t.next_to(c, DOWN)
+                    self.add(d, c, s, t)
+
+        """
+        if isinstance(mobject_or_point, Mobject):
+            mob = mobject_or_point
+            if index_of_submobject_to_align is not None:
+                target_aligner = mob[index_of_submobject_to_align]
+            else:
+                target_aligner = mob
+            target_point = target_aligner.get_bounding_box_point(
+                aligned_edge + direction,
+            )
+        else:
+            target_point = mobject_or_point
+        if submobject_to_align is not None:
+            aligner = submobject_to_align
+        elif index_of_submobject_to_align is not None:
+            aligner = self[index_of_submobject_to_align]
+        else:
+            aligner = self
+        point_to_align = aligner.get_bounding_box_point(aligned_edge - direction)
+        self.shift((target_point - point_to_align + buff * direction) * coor_mask)
+        return self
+
+    def shift_onto_screen(self, **kwargs):
+        space_lengths = [config["frame_x_radius"], config["frame_y_radius"]]
+        for vect in UP, DOWN, LEFT, RIGHT:
+            dim = np.argmax(np.abs(vect))
+            buff = kwargs.get("buff", DEFAULT_MOBJECT_TO_EDGE_BUFFER)
+            max_val = space_lengths[dim] - buff
+            edge_center = self.get_edge_center(vect)
+            if np.dot(edge_center, vect) > max_val:
+                self.to_edge(vect, **kwargs)
+        return self
+
+    def is_off_screen(self):
+        if self.get_left()[0] > config.frame_x_radius:
+            return True
+        if self.get_right()[0] < config.frame_x_radius:
+            return True
+        if self.get_bottom()[1] > config.frame_y_radius:
+            return True
+        if self.get_top()[1] < -config.frame_y_radius:
+            return True
+        return False
+
+    def stretch_about_point(self, factor, dim, point):
+        return self.stretch(factor, dim, about_point=point)
+
+    def rescale_to_fit(self, length, dim, stretch=False, **kwargs):
+        old_length = self.length_over_dim(dim)
+        if old_length == 0:
+            return self
+        if stretch:
+            self.stretch(length / old_length, dim, **kwargs)
+        else:
+            self.scale(length / old_length, **kwargs)
+        return self
+
+    def stretch_to_fit_width(self, width, **kwargs):
+        """Stretches the :class:`~.Mobject` to fit a width, not keeping height/depth proportional.
 
         Returns
         -------
         :class:`Mobject`
             ``self``
+
+        Examples
+        --------
+        ::
+
+            >>> from manim import *
+            >>> sq = Square()
+            >>> sq.height
+            2.0
+            >>> sq.stretch_to_fit_width(5)
+            Square
+            >>> sq.width
+            5.0
+            >>> sq.height
+            2.0
         """
-        for m1, m2 in (self, mobject), (mobject, self):
-            if m1.has_no_points() and m2.has_points():
-                m2.push_self_into_submobjects()
+        return self.rescale_to_fit(width, 0, stretch=True, **kwargs)
+
+    def stretch_to_fit_height(self, height, **kwargs):
+        """Stretches the :class:`~.Mobject` to fit a height, not keeping width/height proportional."""
+        return self.rescale_to_fit(height, 1, stretch=True, **kwargs)
+
+    def stretch_to_fit_depth(self, depth, **kwargs):
+        """Stretches the :class:`~.Mobject` to fit a depth, not keeping width/height proportional."""
+        return self.rescale_to_fit(depth, 1, stretch=True, **kwargs)
+
+    def set_width(self, width, stretch=False, **kwargs):
+        """Scales the :class:`~.Mobject` to fit a width while keeping height/depth proportional.
+
+        Returns
+        -------
+        :class:`Mobject`
+            ``self``
+
+        Examples
+        --------
+        ::
+
+            >>> from manim import *
+            >>> sq = Square()
+            >>> sq.height
+            2.0
+            >>> sq.scale_to_fit_width(5)
+            Square
+            >>> sq.width
+            5.0
+            >>> sq.height
+            5.0
+        """
+        return self.rescale_to_fit(width, 0, stretch=stretch, **kwargs)
+
+    scale_to_fit_width = set_width
+
+    def set_height(self, height, stretch=False, **kwargs):
+        """Scales the :class:`~.Mobject` to fit a height while keeping width/depth proportional."""
+        return self.rescale_to_fit(height, 1, stretch=stretch, **kwargs)
+
+    scale_to_fit_height = set_height
+
+    def set_depth(self, depth, stretch=False, **kwargs):
+        """Scales the :class:`~.Mobject` to fit a depth while keeping width/height proportional."""
+        return self.rescale_to_fit(depth, 2, stretch=stretch, **kwargs)
+
+    scale_to_fit_depth = set_depth
+
+    def set_coord(self, value, dim, direction=ORIGIN):
+        curr = self.get_coord(dim, direction)
+        shift_vect = np.zeros(self.dim)
+        shift_vect[dim] = value - curr
+        self.shift(shift_vect)
+        return self
+
+    def set_x(self, x, direction=ORIGIN):
+        """Set x value of the center of the :class:`~.Mobject` (``int`` or ``float``)"""
+        return self.set_coord(x, 0, direction)
+
+    def set_y(self, y, direction=ORIGIN):
+        """Set y value of the center of the :class:`~.Mobject` (``int`` or ``float``)"""
+        return self.set_coord(y, 1, direction)
+
+    def set_z(self, z, direction=ORIGIN):
+        """Set z value of the center of the :class:`~.Mobject` (``int`` or ``float``)"""
+        return self.set_coord(z, 2, direction)
+
+    def space_out_submobjects(self, factor=1.5, **kwargs):
+        self.scale(factor, **kwargs)
+        for submob in self.submobjects:
+            submob.scale(1.0 / factor)
+        return self
+
+    def move_to(
+        self,
+        point_or_mobject,
+        aligned_edge=ORIGIN,
+        coor_mask=np.array([1, 1, 1]),
+    ):
+        """Move center of the :class:`~.Mobject` to certain coordinate."""
+        if isinstance(point_or_mobject, Mobject):
+            target = point_or_mobject.get_bounding_box_point(aligned_edge)
+        else:
+            target = point_or_mobject
+        point_to_align = self.get_bounding_box_point(aligned_edge)
+        self.shift((target - point_to_align) * coor_mask)
+        return self
+
+    def replace(self, mobject, dim_to_match=0, stretch=False):
+        if not mobject.get_num_points() and not mobject.submobjects:
+            self.scale(0)
+            return self
+        if stretch:
+            for i in range(self.dim):
+                self.rescale_to_fit(mobject.length_over_dim(i), i, stretch=True)
+        else:
+            self.rescale_to_fit(
+                mobject.length_over_dim(dim_to_match),
+                dim_to_match,
+                stretch=False,
+            )
+        self.shift(mobject.get_center() - self.get_center())
+        return self
+
+    def surround(
+        self,
+        mobject: "Mobject",
+        dim_to_match: int = 0,
+        stretch: bool = False,
+        buff: float = MED_SMALL_BUFF,
+    ):
+        self.replace(mobject, dim_to_match, stretch)
+        length = mobject.length_over_dim(dim_to_match)
+        self.scale((length + buff) / length)
+        return self
+
+    def put_start_and_end_on(self, start, end):
+        curr_start, curr_end = self.get_start_and_end()
+        curr_vect = curr_end - curr_start
+        if np.all(curr_vect == 0):
+            raise Exception("Cannot position endpoints of closed loop")
+        target_vect = np.array(end) - np.array(start)
+        axis = (
+            normalize(np.cross(curr_vect, target_vect))
+            if np.linalg.norm(np.cross(curr_vect, target_vect)) != 0
+            else OUT
+        )
+        self.scale(
+            np.linalg.norm(target_vect) / np.linalg.norm(curr_vect),
+            about_point=curr_start,
+        )
+        self.rotate(
+            angle_between_vectors(curr_vect, target_vect),
+            about_point=curr_start,
+            axis=axis,
+        )
+        self.shift(start - curr_start)
+        return self
+
+    # Color functions
+
+    def set_rgba_array(self, color=None, opacity=None, name="rgbas", recurse=True):
+        if color is not None:
+            rgbs = np.array([color_to_rgb(c) for c in listify(color)])
+        if opacity is not None:
+            opacities = listify(opacity)
+
+        # Color only
+        if color is not None and opacity is None:
+            for mob in self.get_family(recurse):
+                mob.data[name] = resize_array(mob.data[name], len(rgbs))
+                mob.data[name][:, :3] = rgbs
+
+        # Opacity only
+        if color is None and opacity is not None:
+            for mob in self.get_family(recurse):
+                mob.data[name] = resize_array(mob.data[name], len(opacities))
+                mob.data[name][:, 3] = opacities
+
+        # Color and opacity
+        if color is not None and opacity is not None:
+            rgbas = np.array([[*rgb, o] for rgb, o in zip(*make_even(rgbs, opacities))])
+            for mob in self.get_family(recurse):
+                mob.data[name] = rgbas.copy()
+        return self
+
+    def set_rgba_array_direct(self, rgbas: np.ndarray, name="rgbas", recurse=True):
+        """Directly set rgba data from `rgbas` and optionally do the same recursively
+        with submobjects. This can be used if the `rgbas` have already been generated
+        with the correct shape and simply need to be set.
+
+        Parameters
+        ----------
+        rgbas
+            the rgba to be set as data
+        name
+            the name of the data attribute to be set
+        recurse
+            set to true to recursively apply this method to submobjects
+        """
+        for mob in self.get_family(recurse):
+            mob.data[name] = rgbas.copy()
+
+    def set_color(self, color, opacity=None, recurse=True):
+        self.set_rgba_array(color, opacity, recurse=False)
+        # Recurse to submobjects differently from how set_rgba_array
+        # in case they implement set_color differently
+        if color is not None:
+            self.color = Color(color)
+        if opacity is not None:
+            self.opacity = opacity
+        if recurse:
+            for submob in self.submobjects:
+                submob.set_color(color, recurse=True)
+        return self
+
+    def set_opacity(self, opacity, recurse=True):
+        self.set_rgba_array(color=None, opacity=opacity, recurse=False)
+        if recurse:
+            for submob in self.submobjects:
+                submob.set_opacity(opacity, recurse=True)
+        return self
+
+    def get_color(self):
+        return rgb_to_hex(self.rgbas[0, :3])
+
+    def get_opacity(self):
+        return self.rgbas[0, 3]
+
+    def set_color_by_gradient(self, *colors):
+        self.set_submobject_colors_by_gradient(*colors)
+        return self
+
+    def set_submobject_colors_by_gradient(self, *colors):
+        if len(colors) == 0:
+            raise Exception("Need at least one color")
+        elif len(colors) == 1:
+            return self.set_color(*colors)
+
+        # mobs = self.family_members_with_points()
+        mobs = self.submobjects
+        new_colors = color_gradient(colors, len(mobs))
+
+        for mob, color in zip(mobs, new_colors):
+            mob.set_color(color)
+        return self
+
+    def fade(self, darkness=0.5, recurse=True):
+        self.set_opacity(1.0 - darkness, recurse=recurse)
+
+    def get_gloss(self):
+        return self.gloss
+
+    def set_gloss(self, gloss, recurse=True):
+        for mob in self.get_family(recurse):
+            mob.gloss = gloss
+        return self
+
+    def get_shadow(self):
+        return self.shadow
+
+    def set_shadow(self, shadow, recurse=True):
+        for mob in self.get_family(recurse):
+            mob.shadow = shadow
+        return self
+
+    # Background rectangle
+
+    def add_background_rectangle(
+        self, color: Optional[Colors] = None, opacity: float = 0.75, **kwargs
+    ):
+        # TODO, this does not behave well when the mobject has points,
+        # since it gets displayed on top
+        """Add a BackgroundRectangle as submobject.
+
+        The BackgroundRectangle is added behind other submobjects.
+
+        This can be used to increase the mobjects visibility in front of a noisy background.
+
+        Parameters
+        ----------
+        color
+            The color of the BackgroundRectangle
+        opacity
+            The opacity of the BackgroundRectangle
+        kwargs
+            Additional keyword arguments passed to the BackgroundRectangle constructor
+
+
+        Returns
+        -------
+        :class:`Mobject`
+            ``self``
+
+        See Also
+        --------
+        :meth:`add_to_back`
+        :class:`~.BackgroundRectangle`
+
+        """
+        from ..mobject.shape_matchers import BackgroundRectangle
+
+        self.background_rectangle = BackgroundRectangle(
+            self, color=color, fill_opacity=opacity, **kwargs
+        )
+        self.add_to_back(self.background_rectangle)
+        return self
+
+    def add_background_rectangle_to_submobjects(self, **kwargs):
+        for submobject in self.submobjects:
+            submobject.add_background_rectangle(**kwargs)
+        return self
+
+    def add_background_rectangle_to_family_members_with_points(self, **kwargs):
+        for mob in self.family_members_with_points():
+            mob.add_background_rectangle(**kwargs)
+        return self
+
+    # Getters
+
+    def get_bounding_box_point(self, direction):
+        bb = self.get_bounding_box()
+        indices = (np.sign(direction) + 1).astype(int)
+        return np.array([bb[indices[i]][i] for i in range(3)])
+
+    def get_edge_center(self, direction) -> np.ndarray:
+        """Get edge coordinates for certain direction."""
+        return self.get_bounding_box_point(direction)
+
+    def get_corner(self, direction) -> np.ndarray:
+        """Get corner coordinates for certain direction."""
+        return self.get_bounding_box_point(direction)
+
+    def get_center(self) -> np.ndarray:
+        """Get center coordinates."""
+        return self.get_bounding_box()[1]
+
+    def get_center_of_mass(self):
+        return self.get_all_points().mean(0)
+
+    def get_boundary_point(self, direction):
+        all_points = self.get_all_points()
+        boundary_directions = all_points - self.get_center()
+        norms = np.linalg.norm(boundary_directions, axis=1)
+        boundary_directions /= np.repeat(norms, 3).reshape((len(norms), 3))
+        index = np.argmax(np.dot(boundary_directions, np.array(direction).T))
+        return all_points[index]
+
+    def get_continuous_bounding_box_point(self, direction):
+        dl, center, ur = self.get_bounding_box()
+        corner_vect = ur - center
+        return center + direction / np.max(
+            np.abs(
+                np.true_divide(
+                    direction,
+                    corner_vect,
+                    out=np.zeros(len(direction)),
+                    where=((corner_vect) != 0),
+                ),
+            ),
+        )
+
+    def get_top(self) -> np.ndarray:
+        """Get top coordinates of a box bounding the :class:`~.Mobject`"""
+        return self.get_edge_center(UP)
+
+    def get_bottom(self) -> np.ndarray:
+        """Get bottom coordinates of a box bounding the :class:`~.Mobject`"""
+        return self.get_edge_center(DOWN)
+
+    def get_right(self) -> np.ndarray:
+        """Get right coordinates of a box bounding the :class:`~.Mobject`"""
+        return self.get_edge_center(RIGHT)
+
+    def get_left(self) -> np.ndarray:
+        """Get left coordinates of a box bounding the :class:`~.Mobject`"""
+        return self.get_edge_center(LEFT)
+
+    def get_zenith(self) -> np.ndarray:
+        """Get zenith coordinates of a box bounding a 3D :class:`~.Mobject`."""
+        return self.get_edge_center(OUT)
+
+    def get_nadir(self) -> np.ndarray:
+        """Get nadir (opposite the zenith) coordinates of a box bounding a 3D :class:`~.Mobject`."""
+        return self.get_edge_center(IN)
+
+    def length_over_dim(self, dim):
+        bb = self.get_bounding_box()
+        return abs((bb[2] - bb[0])[dim])
+
+    def get_width(self):
+        """Returns the width of the mobject."""
+        return self.length_over_dim(0)
+
+    def get_height(self):
+        """Returns the height of the mobject."""
+        return self.length_over_dim(1)
+
+    def get_depth(self):
+        """Returns the depth of the mobject."""
+        return self.length_over_dim(2)
+
+    def get_coord(self, dim: int, direction=ORIGIN):
+        """Meant to generalize ``get_x``, ``get_y`` and ``get_z``"""
+        return self.get_bounding_box_point(direction)[dim]
+
+    def get_x(self, direction=ORIGIN) -> np.float64:
+        """Returns x coordinate of the center of the :class:`~.Mobject` as ``float``"""
+        return self.get_coord(0, direction)
+
+    def get_y(self, direction=ORIGIN) -> np.float64:
+        """Returns y coordinate of the center of the :class:`~.Mobject` as ``float``"""
+        return self.get_coord(1, direction)
+
+    def get_z(self, direction=ORIGIN) -> np.float64:
+        """Returns z coordinate of the center of the :class:`~.Mobject` as ``float``"""
+        return self.get_coord(2, direction)
+
+    def get_start(self):
+        """Returns the point, where the stroke that surrounds the :class:`~.Mobject` starts."""
+        self.throw_error_if_no_points()
+        return np.array(self.points[0])
+
+    def get_end(self):
+        """Returns the point, where the stroke that surrounds the :class:`~.Mobject` ends."""
+        self.throw_error_if_no_points()
+        return np.array(self.points[-1])
+
+    def get_start_and_end(self):
+        """Returns starting and ending point of a stroke as a ``tuple``."""
+        return self.get_start(), self.get_end()
+
+    def point_from_proportion(self, alpha):
+        points = self.points
+        i, subalpha = integer_interpolate(0, len(points) - 1, alpha)
+        return interpolate(points[i], points[i + 1], subalpha)
+
+    def pfp(self, alpha):
+        """Abbreviation for point_from_proportion"""
+        return self.point_from_proportion(alpha)
+
+    def get_pieces(self, n_pieces):
+        template = self.copy()
+        template.submobjects = []
+        alphas = np.linspace(0, 1, n_pieces + 1)
+        return OpenGLGroup(
+            *(
+                template.copy().pointwise_become_partial(self, a1, a2)
+                for a1, a2 in zip(alphas[:-1], alphas[1:])
+            )
+        )
+
+    def get_z_index_reference_point(self):
+        # TODO, better place to define default z_index_group?
+        z_index_group = getattr(self, "z_index_group", self)
+        return z_index_group.get_center()
+
+    # Match other mobject properties
+
+    def match_color(self, mobject: "Mobject"):
+        """Match the color with the color of another :class:`~.Mobject`."""
+        return self.set_color(mobject.get_color())
+
+    def match_dim_size(self, mobject: "Mobject", dim, **kwargs):
+        """Match the specified dimension with the dimension of another :class:`~.Mobject`."""
+        return self.rescale_to_fit(mobject.length_over_dim(dim), dim, **kwargs)
+
+    def match_width(self, mobject: "Mobject", **kwargs):
+        """Match the width with the width of another :class:`~.Mobject`."""
+        return self.match_dim_size(mobject, 0, **kwargs)
+
+    def match_height(self, mobject: "Mobject", **kwargs):
+        """Match the height with the height of another :class:`~.Mobject`."""
+        return self.match_dim_size(mobject, 1, **kwargs)
+
+    def match_depth(self, mobject: "Mobject", **kwargs):
+        """Match the depth with the depth of another :class:`~.Mobject`."""
+        return self.match_dim_size(mobject, 2, **kwargs)
+
+    def match_coord(self, mobject: "Mobject", dim, direction=ORIGIN):
+        """Match the coordinates with the coordinates of another :class:`~.Mobject`."""
+        return self.set_coord(
+            mobject.get_coord(dim, direction),
+            dim=dim,
+            direction=direction,
+        )
+
+    def match_x(self, mobject, direction=ORIGIN):
+        """Match x coord. to the x coord. of another :class:`~.Mobject`."""
+        return self.match_coord(mobject, 0, direction)
+
+    def match_y(self, mobject, direction=ORIGIN):
+        """Match y coord. to the x coord. of another :class:`~.Mobject`."""
+        return self.match_coord(mobject, 1, direction)
+
+    def match_z(self, mobject, direction=ORIGIN):
+        """Match z coord. to the x coord. of another :class:`~.Mobject`."""
+        return self.match_coord(mobject, 2, direction)
+
+    def align_to(
+        self,
+        mobject_or_point: Union["Mobject", Sequence[float]],
+        direction=ORIGIN,
+    ):
+        """
+        Examples:
+        mob1.align_to(mob2, UP) moves mob1 vertically so that its
+        top edge lines ups with mob2's top edge.
+
+        mob1.align_to(mob2, alignment_vect = RIGHT) moves mob1
+        horizontally so that it's center is directly above/below
+        the center of mob2
+        """
+        if isinstance(mobject_or_point, Mobject):
+            point = mobject_or_point.get_bounding_box_point(direction)
+        else:
+            point = mobject_or_point
+
+        for dim in range(self.dim):
+            if direction[dim] != 0:
+                self.set_coord(point[dim], dim, direction)
+        return self
+
+    def get_group_class(self):
+        return OpenGLGroup
+
+    # Alignment
+
+    def align_data_and_family(self, mobject):
+        self.align_family(mobject)
+        self.align_data(mobject)
+
+    def align_data(self, mobject):
+        # In case any data arrays get resized when aligned to shader data
+        # self.refresh_shader_data()
+        for mob1, mob2 in zip(self.get_family(), mobject.get_family()):
+            # Separate out how points are treated so that subclasses
+            # can handle that case differently if they choose
+            mob1.align_points(mob2)
+            for key in mob1.data.keys() & mob2.data.keys():
+                if key == "points":
+                    continue
+                arr1 = mob1.data[key]
+                arr2 = mob2.data[key]
+                if len(arr2) > len(arr1):
+                    mob1.data[key] = resize_preserving_order(arr1, len(arr2))
+                elif len(arr1) > len(arr2):
+                    mob2.data[key] = resize_preserving_order(arr2, len(arr1))
+
+    def align_points(self, mobject):
+        max_len = max(self.get_num_points(), mobject.get_num_points())
+        for mob in (self, mobject):
+            mob.resize_points(max_len, resize_func=resize_preserving_order)
+        return self
+
+    def align_family(self, mobject):
+        mob1 = self
+        mob2 = mobject
+        n1 = len(mob1)
+        n2 = len(mob2)
+        if n1 != n2:
+            mob1.add_n_more_submobjects(max(0, n2 - n1))
+            mob2.add_n_more_submobjects(max(0, n1 - n2))
+        # Recurse
+        for sm1, sm2 in zip(mob1.submobjects, mob2.submobjects):
+            sm1.align_family(sm2)
         return self
 
     def push_self_into_submobjects(self):
-        copy = self.copy()
+        copy = self.deepcopy()
         copy.submobjects = []
-        self.reset_points()
+        self.resize_points(0)
         self.add(copy)
         return self
 
     def add_n_more_submobjects(self, n):
         if n == 0:
-            return
+            return self
 
         curr = len(self.submobjects)
         if curr == 0:
             # If empty, simply add n point mobjects
-            self.submobjects = [self.get_point_mobject() for k in range(n)]
-            return
-
+            null_mob = self.copy()
+            null_mob.set_points([self.get_center()])
+            self.submobjects = [null_mob.copy() for k in range(n)]
+            return self
         target = curr + n
-        # TODO, factor this out to utils so as to reuse
-        # with VMobject.insert_n_curves
         repeat_indices = (np.arange(target) * curr) // target
-        split_factors = [sum(repeat_indices == i) for i in range(curr)]
+        split_factors = [(repeat_indices == i).sum() for i in range(curr)]
         new_submobs = []
         for submob, sf in zip(self.submobjects, split_factors):
             new_submobs.append(submob)
             for _ in range(1, sf):
-                new_submobs.append(submob.copy().fade(1))
+                new_submob = submob.copy()
+                # If the submobject is at all transparent, then
+                # make the copy completely transparent
+                if submob.get_opacity() < 1:
+                    new_submob.set_opacity(0)
+                new_submobs.append(new_submob)
         self.submobjects = new_submobs
         return self
 
-    def repeat_submobject(self, submob):
-        return submob.copy()
+    # Interpolate
 
     def interpolate(self, mobject1, mobject2, alpha, path_func=straight_path()):
         """Turns this :class:`~.Mobject` into an interpolation between ``mobject1``
@@ -2572,28 +2323,51 @@ class Mobject:
                     dotL = Dot(color=WHITE)
                     dotL.shift(2 * LEFT)
 
-                    dotMiddle = VMobject().interpolate(dotL, dotR, alpha=0.3)
+                    dotMiddle = OpenGLVMobject().interpolate(dotL, dotR, alpha=0.3)
 
                     self.add(dotL, dotR, dotMiddle)
         """
-        self.points = path_func(mobject1.points, mobject2.points, alpha)
-        self.interpolate_color(mobject1, mobject2, alpha)
+        for key in self.data:
+            if key in self.locked_data_keys:
+                continue
+            if len(self.data[key]) == 0:
+                continue
+            if key not in mobject1.data or key not in mobject2.data:
+                continue
+
+            if key in ("points", "bounding_box"):
+                func = path_func
+            else:
+                func = interpolate
+
+            self.data[key][:] = func(mobject1.data[key], mobject2.data[key], alpha)
+        for key in self.uniforms:
+            self.uniforms[key] = interpolate(
+                mobject1.uniforms[key],
+                mobject2.uniforms[key],
+                alpha,
+            )
         return self
 
-    def interpolate_color(self, mobject1, mobject2, alpha):
-        raise NotImplementedError("Please override in a child class.")
+    def pointwise_become_partial(self, mobject, a, b):
+        """
+        Set points in such a way as to become only
+        part of mobject.
+        Inputs 0 <= a < b <= 1 determine what portion
+        of mobject to become.
+        """
+        pass  # To implement in subclass
 
     def become(
         self,
         mobject: "Mobject",
-        copy_submobjects: bool = True,
         match_height: bool = False,
         match_width: bool = False,
         match_depth: bool = False,
         match_center: bool = False,
         stretch: bool = False,
     ):
-        """Edit points, colors and submobjects to be identical
+        """Edit all data and submobjects to be identical
         to another :class:`~.Mobject`
 
         .. note::
@@ -2643,105 +2417,255 @@ class Mobject:
         if match_center:
             mobject.move_to(self.get_center())
 
-        self.align_data(mobject)
+        self.align_family(mobject)
         for sm1, sm2 in zip(self.get_family(), mobject.get_family()):
-            sm1.points = np.array(sm2.points)
-            sm1.interpolate_color(sm1, sm2, 1)
+            sm1.set_data(sm2.data)
+            sm1.set_uniforms(sm2.uniforms)
+        self.refresh_bounding_box(recurse_down=True)
         return self
 
-    def match_points(self, mobject: "Mobject", copy_submobjects: bool = True):
-        """Edit points, positions, and submobjects to be identical
-        to another :class:`~.Mobject`, while keeping the style unchanged.
+    # Locking data
 
-        Examples
-        --------
-        .. manim:: MatchPointsScene
-
-            class MatchPointsScene(Scene):
-                def construct(self):
-                    circ = Circle(fill_color=RED, fill_opacity=0.8)
-                    square = Square(fill_color=BLUE, fill_opacity=0.2)
-                    self.add(circ)
-                    self.wait(0.5)
-                    self.play(circ.animate.match_points(square))
-                    self.wait(0.5)
+    def lock_data(self, keys):
         """
-        self.align_data(mobject)
-        for sm1, sm2 in zip(self.get_family(), mobject.get_family()):
-            sm1.points = np.array(sm2.points)
+        To speed up some animations, particularly transformations,
+        it can be handy to acknowledge which pieces of data
+        won't change during the animation so that calls to
+        interpolate can skip this, and so that it's not
+        read into the shader_wrapper objects needlessly
+        """
+        if self.has_updaters:
+            return
+        # Be sure shader data has most up to date information
+        self.refresh_shader_data()
+        self.locked_data_keys = set(keys)
+
+    def lock_matching_data(self, mobject1, mobject2):
+        for sm, sm1, sm2 in zip(
+            self.get_family(),
+            mobject1.get_family(),
+            mobject2.get_family(),
+        ):
+            keys = sm.data.keys() & sm1.data.keys() & sm2.data.keys()
+            sm.lock_data(
+                list(
+                    filter(
+                        lambda key: np.all(sm1.data[key] == sm2.data[key]),
+                        keys,
+                    ),
+                ),
+            )
         return self
+
+    def unlock_data(self):
+        for mob in self.get_family():
+            mob.locked_data_keys = set()
+
+    # Operations touching shader uniforms
+
+    def affects_shader_info_id(func):
+        @wraps(func)
+        def wrapper(self):
+            for mob in self.get_family():
+                func(mob)
+                # mob.refresh_shader_wrapper_id()
+            return self
+
+        return wrapper
+
+    @affects_shader_info_id
+    def fix_in_frame(self):
+        self.is_fixed_in_frame = 1.0
+        return self
+
+    @affects_shader_info_id
+    def unfix_from_frame(self):
+        self.is_fixed_in_frame = 0.0
+        return self
+
+    @affects_shader_info_id
+    def apply_depth_test(self):
+        self.depth_test = True
+        return self
+
+    @affects_shader_info_id
+    def deactivate_depth_test(self):
+        self.depth_test = False
+        return self
+
+    # Shader code manipulation
+
+    def replace_shader_code(self, old, new):
+        # TODO, will this work with VMobject structure, given
+        # that it does not simpler return shader_wrappers of
+        # family?
+        for wrapper in self.get_shader_wrapper_list():
+            wrapper.replace_code(old, new)
+        return self
+
+    def set_color_by_code(self, glsl_code):
+        """
+        Takes a snippet of code and inserts it into a
+        context which has the following variables:
+        vec4 color, vec3 point, vec3 unit_normal.
+        The code should change the color variable
+        """
+        self.replace_shader_code("///// INSERT COLOR FUNCTION HERE /////", glsl_code)
+        return self
+
+    def set_color_by_xyz_func(
+        self,
+        glsl_snippet,
+        min_value=-5.0,
+        max_value=5.0,
+        colormap="viridis",
+    ):
+        """
+        Pass in a glsl expression in terms of x, y and z which returns
+        a float.
+        """
+        # TODO, add a version of this which changes the point data instead
+        # of the shader code
+        for char in "xyz":
+            glsl_snippet = glsl_snippet.replace(char, "point." + char)
+        rgb_list = get_colormap_list(colormap)
+        self.set_color_by_code(
+            "color.rgb = float_to_color({}, {}, {}, {});".format(
+                glsl_snippet,
+                float(min_value),
+                float(max_value),
+                get_colormap_code(rgb_list),
+            ),
+        )
+        return self
+
+    # For shader data
+
+    # def refresh_shader_wrapper_id(self):
+    #     self.shader_wrapper.refresh_id()
+    #     return self
+
+    def get_shader_wrapper(self):
+        from ..renderer.shader_wrapper import ShaderWrapper
+
+        self.shader_wrapper = ShaderWrapper(
+            vert_data=self.get_shader_data(),
+            vert_indices=self.get_shader_vert_indices(),
+            uniforms=self.get_shader_uniforms(),
+            depth_test=self.depth_test,
+            texture_paths=self.texture_paths,
+            render_primitive=self.render_primitive,
+            shader_folder=self.__class__.shader_folder,
+        )
+        return self.shader_wrapper
+
+    def get_shader_wrapper_list(self):
+        shader_wrappers = it.chain(
+            [self.get_shader_wrapper()],
+            *(sm.get_shader_wrapper_list() for sm in self.submobjects),
+        )
+        batches = batch_by_property(shader_wrappers, lambda sw: sw.get_id())
+
+        result = []
+        for wrapper_group, _ in batches:
+            shader_wrapper = wrapper_group[0]
+            if not shader_wrapper.is_valid():
+                continue
+            shader_wrapper.combine_with(*wrapper_group[1:])
+            if len(shader_wrapper.vert_data) > 0:
+                result.append(shader_wrapper)
+        return result
+
+    def check_data_alignment(self, array, data_key):
+        # Makes sure that self.data[key] can be broadcast into
+        # the given array, meaning its length has to be either 1
+        # or the length of the array
+        d_len = len(self.data[data_key])
+        if d_len != 1 and d_len != len(array):
+            self.data[data_key] = resize_with_interpolation(
+                self.data[data_key],
+                len(array),
+            )
+        return self
+
+    def get_resized_shader_data_array(self, length):
+        # If possible, try to populate an existing array, rather
+        # than recreating it each frame
+        points = self.points
+        shader_data = np.zeros(len(points), dtype=self.shader_dtype)
+        return shader_data
+
+    def read_data_to_shader(self, shader_data, shader_data_key, data_key):
+        if data_key in self.locked_data_keys:
+            return
+        self.check_data_alignment(shader_data, data_key)
+        shader_data[shader_data_key] = self.data[data_key]
+
+    def get_shader_data(self):
+        shader_data = self.get_resized_shader_data_array(self.get_num_points())
+        self.read_data_to_shader(shader_data, "point", "points")
+        return shader_data
+
+    def refresh_shader_data(self):
+        self.get_shader_data()
+
+    def get_shader_uniforms(self):
+        return self.uniforms
+
+    def get_shader_vert_indices(self):
+        return self.shader_indices
+
+    @property
+    def submobjects(self):
+        return self._submobjects
+
+    @submobjects.setter
+    def submobjects(self, submobject_list):
+        self.remove(*self.submobjects)
+        self.add(*submobject_list)
 
     # Errors
+
     def throw_error_if_no_points(self):
-        if self.has_no_points():
-            caller_name = sys._getframe(1).f_code.co_name
-            raise Exception(
-                f"Cannot call Mobject.{caller_name} for a Mobject with no points",
+        if not self.has_points():
+            message = (
+                "Cannot call Mobject.{} " + "for a Mobject with no points"
             )
-
-    # About z-index
-    def set_z_index(
-        self,
-        z_index_value: float,
-        family: bool = True,
-    ) -> "VMobject":
-        """Sets the :class:`~.Mobject`'s :attr:`z_index` to the value specified in `z_index_value`.
-
-        Parameters
-        ----------
-        z_index_value
-            The new value of :attr:`z_index` set.
-        family
-            If ``True``, the :attr:`z_index` value of all submobjects is also set.
-
-        Returns
-        -------
-        :class:`Mobject`
-            The Mobject itself, after :attr:`z_index` is set. For chaining purposes. (Returns `self`.)
-
-        Examples
-        --------
-        .. manim:: SetZIndex
-            :save_last_frame:
-
-            class SetZIndex(Scene):
-                def construct(self):
-                    text = Text('z_index = 3', color = PURE_RED).shift(UP).set_z_index(3)
-                    square = Square(2, fill_opacity=1).set_z_index(2)
-                    tex = Tex(r'zIndex = 1', color = PURE_BLUE).shift(DOWN).set_z_index(1)
-                    circle = Circle(radius = 1.7, color = GREEN, fill_opacity = 1) # z_index = 0
-
-                    # Displaying order is now defined by z_index values
-                    self.add(text)
-                    self.add(square)
-                    self.add(tex)
-                    self.add(circle)
-        """
-        if family:
-            for submob in self.submobjects:
-                submob.set_z_index(z_index_value, family=family)
-        self.z_index = z_index_value
-        return self
-
-    def set_z_index_by_z_coordinate(self):
-        """Sets the :class:`~.Mobject`'s z coordinate to the value of :attr:`z_index`.
-
-        Returns
-        -------
-        :class:`Mobject`
-            The Mobject itself, after :attr:`z_index` is set. (Returns `self`.)
-        """
-        z_coord = self.get_center()[-1]
-        self.set_z_index(z_coord)
-        return self
+            caller_name = sys._getframe(1).f_code.co_name
+            raise Exception(message.format(caller_name))
 
 
-class Group(Mobject, metaclass=ConvertToOpenGL):
-    """Groups together multiple :class:`Mobjects <.Mobject>`."""
-
+class OpenGLGroup(Mobject):
     def __init__(self, *mobjects, **kwargs):
+        if not all([isinstance(m, Mobject) for m in mobjects]):
+            raise Exception("All submobjects must be of type Mobject")
         super().__init__(**kwargs)
         self.add(*mobjects)
+
+
+class OpenGLPoint(Mobject):
+    def __init__(
+        self, location=ORIGIN, artificial_width=1e-6, artificial_height=1e-6, **kwargs
+    ):
+        self.artificial_width = artificial_width
+        self.artificial_height = artificial_height
+        super().__init__(**kwargs)
+        self.set_location(location)
+
+    def get_width(self):
+        return self.artificial_width
+
+    def get_height(self):
+        return self.artificial_height
+
+    def get_location(self):
+        return self.points[0].copy()
+
+    def get_bounding_box_point(self, *args, **kwargs):
+        return self.get_location()
+
+    def set_location(self, new_loc):
+        self.set_points(np.array(new_loc, ndmin=2, dtype=float))
 
 
 class _AnimationBuilder:
